TES: Improved organization of test methods.
