+++ pysatMagVect/tests/test_core.py
@@ -27,57 +27,9 @@ else:
     dc = None
     dview = None
 
-# results from omniweb calculator
-omni_list = [[550. , 20.00   , 0.00 , 29.77,  359.31,  -9.04 ,   3.09],
-    [550. , 20.00   , 7.50 , 29.50  ,  7.19 , -8.06 ,   9.54],
-    [550. , 20.00   ,15.00 , 28.34  , 15.01 , -7.51 ,  16.20],
-    [550. , 20.00   ,22.50 , 27.61  , 22.68 , -7.75 ,  23.10],
-    [550. , 20.00   ,30.00 , 27.36  , 30.27 , -8.85 ,  30.34],
-    [550. , 20.00   ,37.50 , 27.10  , 37.79 ,-10.13 ,  38.00],
-   [550. , 20.00   ,45.00  ,26.89   ,45.24 ,-11.24  , 46.09],
-   [ 550.  ,20.00   ,52.50 , 26.77  , 52.65 ,-11.80 ,  54.31],
-   [550.  ,20.00   ,60.00  ,26.75   ,60.05 ,-11.77  , 62.31],
-   [550.  ,20.00   ,67.50  ,26.80   ,67.49 ,-11.34  , 69.94],
-   [550.  ,20.00   ,75.00  ,26.89   ,74.95 ,-10.80  , 77.22],
-   [550.  ,20.00   ,82.50  ,26.97   ,82.45 ,-10.37  , 84.27],
-   [550.  ,20.00   ,90.00  ,27.01   ,89.94 ,-10.20  , 91.22],
-   [550.  ,20.00   ,97.50  ,27.01   ,97.42 ,-10.31  , 98.19],
-   [550.  ,20.00  ,105.00  ,26.99  ,104.86 ,-10.64  ,105.27],
-   [550.  ,20.00  ,112.50  ,27.02  ,112.26 ,-11.00  ,112.49],
-   [550.  ,20.00  ,120.00  ,27.11  ,119.65 ,-11.25  ,119.77],
-   [550.  ,20.00  ,127.50  ,27.28  ,127.08 ,-11.35  ,126.99],
-   [550.  ,20.00  ,135.00  ,27.50  ,134.59 ,-11.41  ,134.03],
-   [550.  ,20.00  ,142.50  ,27.69  ,142.23 ,-11.63  ,140.83],
-   [550.  ,20.00  ,150.00  ,27.76  ,149.98 ,-12.20  ,147.44],
-   [550.  ,20.00  ,157.50  ,27.66  ,157.81 ,-13.19  ,153.90],
-   [550.  ,20.00  ,165.00  ,27.37  ,165.64 ,-14.60  ,160.32],
-   [550.  ,20.00  ,172.50  ,26.96  ,173.39 ,-16.33  ,166.77],
-   [550.  ,20.00  ,180.00  ,26.50  ,181.04 ,-18.23  ,173.34],
-   [550.  ,20.00  ,187.50  ,26.10  ,188.60 ,-20.15  ,180.05],
-   [550.  ,20.00  ,195.00  ,25.78  ,196.11 ,-22.00  ,186.89],
-   [550.  ,20.00  ,202.50  ,25.53  ,203.62 ,-23.77  ,193.77],
-   [550.  ,20.00  ,210.00  ,25.31  ,211.12 ,-25.52  ,200.61],
-   [550.  ,20.00  ,217.50  ,25.09  ,218.62 ,-27.37  ,207.40],
-   [550.  ,20.00  ,225.00  ,24.87  ,226.09 ,-29.37  ,214.17],
-   [550.  ,20.00  ,232.50  ,24.64  ,233.52 ,-31.56  ,220.97],
-   [550.  ,20.00  ,240.00  ,24.42  ,240.93 ,-33.92  ,227.85],
-   [550.  ,20.00  ,247.50  ,24.19  ,248.29 ,-36.49  ,234.86],
-   [550.  ,20.00  ,255.00  ,23.98  ,255.62 ,-39.28  ,242.16],
-   [550.  ,20.00  ,262.50  ,23.80  ,262.90 ,-42.28  ,250.04],
-   [550.  ,20.00  ,270.00  ,23.66  ,270.13 ,-45.35  ,259.07],
-   [550.  ,20.00  ,277.50  ,23.61  ,277.33 ,-48.05 , 270.08],
-   [550.  ,20.00  ,285.00  ,23.65  ,284.50 ,-49.36 , 283.83],
-   [550.  ,20.00  ,292.50  ,23.81  ,291.67 ,-47.58 , 299.53],
-   [550.  ,20.00  ,300.00  ,24.10  ,298.85 ,-41.86 , 313.55],
-   [550.  ,20.00  ,307.50  ,24.55  ,306.06 ,-34.06 , 323.42],
-   [550.  ,20.00  ,315.00  ,25.14  ,313.34 ,-26.36 , 330.13],
-   [550.  ,20.00  ,322.50  ,25.87  ,320.73 ,-19.73 , 335.31],
-   [550.  ,20.00  ,330.00  ,26.63  ,328.27 ,-14.56 , 340.08],
-   [550.  ,20.00  ,337.50  ,28.33  ,335.63 ,-12.03 , 345.55],
-   [550.  ,20.00  ,345.00  ,29.45  ,343.37 ,-10.82 , 351.18],
-   [550.  ,20.00  ,352.50  ,30.17  ,351.27 , -9.90 , 356.93]]
-omni = pds.DataFrame(omni_list, columns=['p_alt', 'p_lat', 'p_long', 'n_lat', 'n_long', 's_lat', 's_long'])
 
+
+# Methods to generate data sets used by testing routines.
 def gen_data_fixed_alt(alt):
     """Generate data between -90 and 90 degrees latitude, almost."""
     import itertools
@@ -149,7 +101,12 @@ def gen_plot_grid_fixed_alt(alt):
     return lat_dim, long_dim, alt_dim
 
 
-class TestCore():
+
+
+################## UNIT VECTOR TESTS ###############################
+
+
+class TestUnitVectors():
 
     def __init__(self):
         # placeholder for data management features
@@ -160,351 +117,6 @@ class TestCore():
         self.dc = dc
 
         return
-
-    def test_field_line_tracing_against_vitmo(self):
-        """Compare model to http://omniweb.gsfc.nasa.gov/vitmo/cgm_vitmo.html"""
-
-        # convert position to ECEF
-        ecf_x,ecf_y,ecf_z = pymv.geocentric_to_ecef(omni['p_lat'],
-                                                  omni['p_long'],
-                                                  omni['p_alt'])
-        trace_n = []
-        trace_s = []
-        date = datetime.datetime(2000, 1, 1)
-        for x,y,z in zip(ecf_x,ecf_y,ecf_z):
-            # trace north and south, take last points
-            trace_n.append(pymv.field_line_trace(np.array([x,y,z]), date, 1., 0., step_size=0.5, max_steps=1.E6)[-1,:])
-            trace_s.append(pymv.field_line_trace(np.array([x,y,z]), date, -1., 0., step_size=0.5, max_steps=1.E6)[-1,:])
-        trace_n = pds.DataFrame(trace_n, columns = ['x', 'y', 'z'])
-        trace_n['lat'], trace_n['long'], trace_n['altitude'] = pymv.ecef_to_geocentric(trace_n['x'], trace_n['y'], trace_n['z'])
-        trace_s = pds.DataFrame(trace_s, columns = ['x', 'y', 'z'])
-        trace_s['lat'], trace_s['long'], trace_s['altitude'] = pymv.ecef_to_geocentric(trace_s['x'], trace_s['y'], trace_s['z'])
-
-        # ensure longitudes are all 0-360
-        idx, = np.where(omni['n_long'] < 0)
-        omni.ix[idx, 'n_long'] += 360.
-        idx, = np.where(omni['s_long'] < 0)
-        omni.ix[idx, 's_long'] += 360.
-
-        idx, = np.where(trace_n['long'] < 0)
-        trace_n.ix[idx, 'long'] += 360.
-        idx, = np.where(trace_s['long'] < 0)
-        trace_s.ix[idx, 'long'] += 360.
-
-        # compute difference between OMNI and local calculation
-        # there is a difference near 0 longitude, ignore this area
-        diff_n_lat = (omni['n_lat'] - trace_n['lat'])[4:-4]
-        diff_n_lon = (omni['n_long'] - trace_n['long'])[4:-4]
-        diff_s_lat = (omni['s_lat'] - trace_s['lat'])[4:-4]
-        diff_s_lon = (omni['s_long'] - trace_s['long'])[4:-4]
-
-        try:
-            f = plt.figure()
-            plt.plot(omni['n_long'], omni['n_lat'], 'r.', label='omni')
-            plt.plot(omni['s_long'], omni['s_lat'], 'r.', label='_omni')
-            plt.plot(trace_n['long'], trace_n['lat'], 'b.', label='UTD')
-            plt.plot(trace_s['long'], trace_s['lat'], 'b.', label='_UTD')
-            plt.title('Comparison of Magnetic Footpoints for Field Lines through 20 Lat, 550 km')
-            plt.xlabel('Geographic Longitude')
-            plt.ylabel('Geographic Latitude')
-            plt.legend(loc=0)
-            plt.xlim((0,360.))
-            plt.savefig('magnetic_footpoint_comparison.pdf')
-            print('Saving magnetic_footpoint_comparison.pdf')
-        except:
-            pass
-
-        # better than 0.5 km accuracy expected for settings above
-        assert np.all(np.nanstd(diff_n_lat) < .5)
-        assert np.all(np.nanstd(diff_n_lon) < .5)
-        assert np.all(np.nanstd(diff_s_lat) < .5)
-        assert np.all(np.nanstd(diff_s_lon) < .5)
-
-
-
-    def test_geodetic_to_ecef_to_geodetic(self):
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        lat, elong, alt = pymv.ecef_to_geodetic(ecf_x, ecf_y, ecf_z)
-        lat2, elong2, alt2 = pymv.python_ecef_to_geodetic(ecf_x, ecf_y, ecf_z)
-
-        idx, = np.where(elong < 0)
-        elong[idx] += 360.
-
-        idx, = np.where(elong2 < 0)
-        elong2[idx] += 360.
-
-        d_lat = lat - lats
-        d_long = elong - longs
-        d_alt = alt - alts
-
-        assert np.all(np.abs(d_lat) < 1.E-5)
-        assert np.all(np.abs(d_long) < 1.E-5)
-        assert np.all(np.abs(d_alt) < 1.E-5)
-
-        d_lat = lat2 - lats
-        d_long = elong2 - longs
-        d_alt = alt2 - alts
-
-        assert np.all(np.abs(d_lat) < 1.E-5)
-        assert np.all(np.abs(d_long) < 1.E-5)
-        assert np.all(np.abs(d_alt) < 1.E-5)
-
-    def test_geodetic_to_ecef_to_geodetic_via_different_methods(self):
-
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        methods = ['closed', 'iterative']
-        for method in methods:
-            lat, elong, alt = pymv.python_ecef_to_geodetic(ecf_x, ecf_y, ecf_z,
-                                                  method=method)
-
-            idx, = np.where(elong < 0)
-            elong[idx] += 360.
-
-            d_lat = lat - lats
-            d_long = elong - longs
-            d_alt = alt - alts
-
-            assert np.all(np.abs(d_lat) < 1.E-5)
-            assert np.all(np.abs(d_long) < 1.E-5)
-            assert np.all(np.abs(d_alt) < 1.E-5)
-
-
-    def test_geodetic_to_ecef_to_geocentric_to_ecef_to_geodetic(self):
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        geo_lat, geo_long, geo_alt = pymv.ecef_to_geocentric(ecf_x, ecf_y, ecf_z)
-
-        ecfs_x,ecfs_y,ecfs_z = pymv.geocentric_to_ecef(geo_lat, geo_long, geo_alt)
-
-        lat, elong, alt = pymv.ecef_to_geodetic(ecfs_x, ecfs_y, ecfs_z)
-
-        idx, = np.where(elong < 0)
-        elong[idx] += 360.
-
-        d_lat = lat - lats
-        d_long = elong - longs
-        d_alt = alt - alts
-
-        assert np.all(np.abs(d_lat) < 1.E-5)
-        assert np.all(np.abs(d_long) < 1.E-5)
-        assert np.all(np.abs(d_alt) < 1.E-5)
-        assert np.all(np.abs(ecf_x-ecfs_x) < 1.E-5)
-        assert np.all(np.abs(ecf_y-ecfs_y) < 1.E-5)
-        assert np.all(np.abs(ecf_z-ecfs_z) < 1.E-5)
-
-
-    def test_geocentric_to_ecef_to_geocentric(self):
-
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geocentric_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        lat, elong, alt = pymv.ecef_to_geocentric(ecf_x, ecf_y, ecf_z)
-
-        idx, = np.where(elong < 0)
-        elong[idx] += 360.
-
-        d_lat = lat - lats
-        d_long = elong - longs
-        d_alt = alt - alts
-
-        assert np.all(np.abs(d_lat) < 1.E-5)
-        assert np.all(np.abs(d_long) < 1.E-5)
-        assert np.all(np.abs(d_alt) < 1.E-5)
-
-
-
-    def test_tracing_accuracy(self):
-
-        lats, longs, alts = gen_trace_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        # step size to be tried, move by inverse powers of 2 (half step size)
-        steps_goal = np.arange(13)
-        steps_goal = 1000./2**steps_goal
-        # max number of steps (fixed)
-        max_steps_goal = steps_goal*0+1E6
-
-        date = datetime.datetime(2000, 1, 1)
-        dx = []
-        dy = []
-        dz = []
-
-        # set up multi
-        # if self.dc is not None:
-        if False:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for x, y, z in zip(ecf_x, ecf_y, ecf_z):
-                for steps, max_steps in zip(steps_goal, max_steps_goal):
-                    # iterate through target cyclicly and run commands
-                    dview.targets = targets.next()
-                    pending.append(dview.apply_async(pymv.field_line_trace, np.array([x,y,z]), date, 1., 0.,
-                                                    step_size=steps,
-                                                    max_steps=max_steps))
-            # for x, y, z in zip(ecf_x, ecf_y, ecf_z):
-                out = []
-                for steps, max_steps in zip(steps_goal, max_steps_goal):
-                    # collect output
-                    trace_n = pending.pop(0).get()
-                    pt = trace_n[-1,:]
-                    out.append(pt)
-
-                final_pt = pds.DataFrame(out, columns = ['x', 'y', 'z'])
-                dx.append(np.abs(final_pt.ix[1:, 'x'].values - final_pt.ix[:,'x'].values[:-1]))
-                dy.append(np.abs(final_pt.ix[1:, 'y'].values - final_pt.ix[:,'y'].values[:-1]))
-                dz.append(np.abs(final_pt.ix[1:, 'z'].values- final_pt.ix[:,'z'].values[:-1]))
-        else:
-            for x, y, z in zip(ecf_x, ecf_y, ecf_z):
-                out = []
-                for steps, max_steps in zip(steps_goal, max_steps_goal):
-                    trace_n = pymv.field_line_trace(np.array([x,y,z]), date, 1., 0.,
-                                                    step_size=steps,
-                                                    max_steps=max_steps)
-                    pt = trace_n[-1,:]
-                    out.append(pt)
-
-                final_pt = pds.DataFrame(out, columns = ['x', 'y', 'z'])
-                dx.append(np.abs(final_pt.ix[1:, 'x'].values - final_pt.ix[:,'x'].values[:-1]))
-                dy.append(np.abs(final_pt.ix[1:, 'y'].values - final_pt.ix[:,'y'].values[:-1]))
-                dz.append(np.abs(final_pt.ix[1:, 'z'].values- final_pt.ix[:,'z'].values[:-1]))
-
-
-        dx = pds.DataFrame(dx)
-        dy = pds.DataFrame(dy)
-        dz = pds.DataFrame(dz)
-
-        try:
-            plt.figure()
-            yerrx = np.nanstd(np.log10(dx), axis=0)
-            yerry = np.nanstd(np.log10(dy), axis=0)
-            yerrz = np.nanstd(np.log10(dz), axis=0)
-            plt.errorbar(np.log10(steps_goal[1:]), np.log10(dx.mean(axis=0)),
-                          yerr=yerrx, label='x')
-            plt.errorbar(np.log10(steps_goal[1:]), np.log10(dy.mean(axis=0)),
-                        yerr=yerry, label='y')
-            plt.errorbar(np.log10(steps_goal[1:]), np.log10(dz.mean(axis=0)),
-                        yerr=yerrz, label='z')
-            plt.xlabel('Log Step Size (km)')
-            plt.ylabel('Change in Foot Point Position (km)')
-            plt.title("Change in Final ECEF Position")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('Footpoint_position_vs_step_size.pdf' )
-            plt.close()
-        except:
-            pass
-
-    def test_apex_info_accuracy(self):
-
-        lats, longs, alts = gen_trace_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        # step size to be tried
-        fine_steps_goal = np.array([25.6, 12.8, 6.4, 3.2, 1.6, 0.8, 0.4, 0.2,
-                                    0.1, 0.05, .025, .0125, .00625, .003125,
-                                    .0015625, .00078125, .000390625, .0001953125,
-                                    .0001953125/2., .0001953125/4., .0001953125/8.,
-                                    .0001953125/16., .0001953125/32., .0001953125/64.,
-                                    .0001953125/128., .0001953125/256., .0001953125/512.,
-                                    .0001953125/1024., .0001953125/2048., .0001953125/4096.])
-
-        date = datetime.datetime(2000, 1, 1)
-        dx = []
-        dy = []
-        dz = []
-        dh = []
-
-        # set up multi
-        # if self.dc is not None:
-        if False:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for lat, lon, alt in zip(lats, longs, alts):
-                for steps in fine_steps_goal:
-                    # iterate through target cyclicly and run commands
-                    dview.targets = targets.next()
-                    pending.append(dview.apply_async(pymv.apex_location_info, [lat],
-                                                     [lon], [alt], [date], fine_step_size=steps,
-                                                     return_geodetic=True))
-            # for x, y, z in zip(ecf_x, ecf_y, ecf_z):
-                out = []
-                for steps in fine_steps_goal:
-                    # collect output
-                    x, y, z, _, _, apex_height = pending.pop(0).get()
-                    pt = [x[0], y[0], z[0], apex_height[0]]
-                    out.append(pt)
-
-                final_pt = pds.DataFrame(out, columns = ['x', 'y', 'z', 'h'])
-                dx.append(np.abs(final_pt.ix[1:, 'x'].values - final_pt.ix[:,'x'].values[:-1]))
-                dy.append(np.abs(final_pt.ix[1:, 'y'].values - final_pt.ix[:,'y'].values[:-1]))
-                dz.append(np.abs(final_pt.ix[1:, 'z'].values - final_pt.ix[:,'z'].values[:-1]))
-                dh.append(np.abs(final_pt.ix[1:, 'h'].values - final_pt.ix[:,'h'].values[:-1]))
-        else:
-            for lat, lon, alt in zip(lats, longs, alts):
-                out = []
-                for steps in fine_steps_goal:
-                    x, y, z, _, _, apex_height = pymv.apex_location_info([lat], [lon], [alt], [date], fine_step_size=steps,
-                                                                         return_geodetic=True)
-                    pt = [x[0], y[0], z[0], apex_height[0]]
-                    out.append(pt)
-
-                final_pt = pds.DataFrame(out, columns = ['x', 'y', 'z', 'h'])
-                dx.append(np.abs(final_pt.ix[1:, 'x'].values - final_pt.ix[:,'x'].values[:-1]))
-                dy.append(np.abs(final_pt.ix[1:, 'y'].values - final_pt.ix[:,'y'].values[:-1]))
-                dz.append(np.abs(final_pt.ix[1:, 'z'].values - final_pt.ix[:,'z'].values[:-1]))
-                dh.append(np.abs(final_pt.ix[1:, 'h'].values - final_pt.ix[:,'h'].values[:-1]))
-
-
-        dx = pds.DataFrame(dx)
-        dy = pds.DataFrame(dy)
-        dz = pds.DataFrame(dz)
-        dh = pds.DataFrame(dh)
-
-        try:
-
-
-            plt.figure()
-            yerrx = np.nanstd(np.log10(dx), axis=0)
-            yerry = np.nanstd(np.log10(dy), axis=0)
-            yerrz = np.nanstd(np.log10(dz), axis=0)
-            yerrh = np.nanstd(np.log10(dh), axis=0)
-
-            plt.errorbar(np.log10(fine_steps_goal[1:]), np.log10(dx.mean(axis=0)),
-                            yerr=yerrx,
-                            label='x')
-            plt.errorbar(np.log10(fine_steps_goal[1:]), np.log10(dy.mean(axis=0)),
-                            yerr=yerry,
-                            label='y')
-            plt.errorbar(np.log10(fine_steps_goal[1:]), np.log10(dz.mean(axis=0)),
-                            yerr=yerrz,
-                            label='z')
-            plt.errorbar(np.log10(fine_steps_goal[1:]), np.log10(dh.mean(axis=0)),
-                            yerr=yerrh,
-                            label='h')
-
-            plt.xlabel('Log Step Size (km)')
-            plt.ylabel('Change in Apex Position (km)')
-            plt.title("Change in Field Apex Position vs Fine Step Size")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('apex_location_vs_step_size.pdf' )
-            plt.close()
-        except:
-            pass
-
 
 
     def test_unit_vector_step_size_sensitivity(self):
@@ -523,8 +135,7 @@ class TestCore():
         dmz = []
 
         # set up multi
-        # if self.dc is not None:
-        if False:
+        if self.dc is not None:
             import itertools
             targets = itertools.cycle(dc.ids)
             pending = []
@@ -622,253 +233,8 @@ class TestCore():
         except:
             pass
 
-    def test_tracing_accuracy_w_recursion(self):
-
-        lats, longs, alts = gen_trace_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        # step size to be tried
-        steps_goal = np.array([5.]*13)
-        # max number of steps (fixed)
-        max_steps_goal = np.arange(13)
-        max_steps_goal = 100000./2**max_steps_goal
-
-        date = datetime.datetime(2000, 1, 1)
-        dx = []
-        dy = []
-        dz = []
-        for x, y, z in zip(ecf_x, ecf_y, ecf_z):
-            out = []
-            for steps, max_steps in zip(steps_goal, max_steps_goal):
-                trace_n = pymv.field_line_trace(np.array([x,y,z]), date, 1., 0.,
-                                                step_size=steps,
-                                                max_steps=max_steps)
-                pt = trace_n[-1,:]
-                out.append(pt)
-
-            final_pt = pds.DataFrame(out, columns = ['x', 'y', 'z'])
-            dx.append(np.abs(final_pt.ix[1:, 'x'].values - final_pt.ix[:,'x'].values[:-1]))
-            dy.append(np.abs(final_pt.ix[1:, 'y'].values - final_pt.ix[:,'y'].values[:-1]))
-            dz.append(np.abs(final_pt.ix[1:, 'z'].values - final_pt.ix[:,'z'].values[:-1]))
-        dx = pds.DataFrame(dx)
-        dy = pds.DataFrame(dy)
-        dz = pds.DataFrame(dz)
-
-        try:
-            plt.figure()
-            yerrx = np.nanstd(np.log10(dx), axis=0)
-            yerry = np.nanstd(np.log10(dy), axis=0)
-            yerrz = np.nanstd(np.log10(dz), axis=0)
-            plt.errorbar(np.log10(max_steps_goal[1:]), np.log10(dx.mean(axis=0)),
-                          yerr=yerrx,
-                          label='x')
-            plt.errorbar(np.log10(max_steps_goal[1:]), np.log10(dy.mean(axis=0)),
-                          yerr=yerry,
-                          label='y')
-            plt.errorbar(np.log10(max_steps_goal[1:]), np.log10(dz.mean(axis=0)),
-                          yerr=yerrz,
-                          label='z')
-            plt.xlabel('Log Number of Steps per Run')
-            plt.ylabel('Change in Foot Point Position (km)')
-            plt.title("Change in Final ECEF Position, Recursive Calls")
-            plt.legend()
-            plt.tight_layout()
-            plt.ylabel('Log Change in Foot Point Position (km)')
-            plt.savefig('Footpoint_position_vs_max_steps_recursion.pdf' )
-            plt.close()
-        except:
-            pass
-
-    def test_tracing_accuracy_w_recursion_step_size(self):
-        lats, longs, alts = gen_trace_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        # step size to be tried
-        steps_goal = np.arange(13)
-        steps_goal = 500./2**steps_goal
-        # max number of steps (fixed)
-        max_steps_goal = np.array([10000.]*13)
-
-        date = datetime.datetime(2000, 1, 1)
-        dx = []
-        dy = []
-        dz = []
-        for x, y, z in zip(ecf_x, ecf_y, ecf_z):
-            out = []
-            for steps, max_steps in zip(steps_goal, max_steps_goal):
-                trace_n = pymv.field_line_trace(np.array([x,y,z]), date, 1., 0.,
-                                                step_size=steps,
-                                                max_steps=max_steps)
-                pt = trace_n[-1,:]
-                out.append(pt)
-
-            final_pt = pds.DataFrame(out, columns = ['x', 'y', 'z'])
-            dx.append(np.abs(final_pt.ix[1:, 'x'].values - final_pt.ix[:,'x'].values[:-1]))
-            dy.append(np.abs(final_pt.ix[1:, 'y'].values - final_pt.ix[:,'y'].values[:-1]))
-            dz.append(np.abs(final_pt.ix[1:, 'z'].values- final_pt.ix[:,'z'].values[:-1]))
-        dx = pds.DataFrame(dx)
-        dy = pds.DataFrame(dy)
-        dz = pds.DataFrame(dz)
-
-        try:
-            plt.figure()
-            yerrx = np.nanstd(np.log10(dx), axis=0)
-            yerry = np.nanstd(np.log10(dy), axis=0)
-            yerrz = np.nanstd(np.log10(dz), axis=0)
-            plt.errorbar(np.log10(steps_goal[:-1]), np.log10(dx.mean(axis=0)),
-                          yerr=yerrx,
-                          label='x')
-            plt.errorbar(np.log10(steps_goal[:-1]), np.log10(dy.mean(axis=0)),
-                          yerr=yerry,
-                          label='y')
-            plt.errorbar(np.log10(steps_goal[:-1]), np.log10(dz.mean(axis=0)),
-                          yerr=yerrz,
-                          label='z')
-            plt.xlabel('Log Step Size (km)')
-            plt.ylabel('Log Change in Foot Point Position (km)')
-            plt.title("Change in Final ECEF Position, Recursive Calls")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('Footpoint_position_vs_step_size__recursion.pdf' )
-            plt.close()
-        except:
-            pass
 
 
-    def test_unit_vector_plots(self):
-        import matplotlib.pyplot as plt
-        from mpl_toolkits.mplot3d import Axes3D
-        import os
-        on_travis = os.environ.get('ONTRAVIS') == 'True'
-
-        # convert OMNI position to ECEF
-        p_long = np.arange(0.,360.,12.)
-        p_alt = 0*p_long + 550.
-        p_lats = [ 5., 10., 15., 20., 25., 30.]
-
-        truthiness = []
-        for i,p_lat in enumerate(p_lats):
-
-            trace_s = []
-            if not on_travis:
-                fig = plt.figure()
-                ax = fig.add_subplot(111, projection='3d')
-
-            #
-            date = datetime.datetime(2000,1,1)
-            ecef_x,ecef_y,ecef_z = pymv.geocentric_to_ecef(p_lat, p_long, p_alt)
-
-            for j,(x,y,z) in enumerate(zip(ecef_x, ecef_y, ecef_z)):
-                # perform field line traces
-                trace_n = pymv.field_line_trace(np.array([x,y,z]), date, 1., 0., step_size=.5, max_steps=1.E6)
-                trace_s = pymv.field_line_trace(np.array([x,y,z]), date, -1., 0., step_size=.5, max_steps=1.E6)
-                # combine together, S/C position is first for both
-                # reverse first array and join so plotting makes sense
-                trace = np.vstack((trace_n[::-1], trace_s))
-                trace = pds.DataFrame(trace, columns=['x','y','z'])
-                # plot field-line
-                if not on_travis:
-                    ax.plot(trace['x'],trace['y'],trace['z'] , 'b')
-                    plt.xlabel('X')
-                    plt.ylabel('Y')
-                    ax.set_zlabel('Z')
-                # clear stored data
-                self.inst.data = pds.DataFrame()
-                # downselect, reduce number of points
-                trace = trace.ix[::1000,:]
-
-                # compute magnetic field vectors
-                # need to provide alt, latitude, and longitude in geodetic coords
-                latitude, longitude, altitude = pymv.ecef_to_geodetic(trace['x'], trace['y'], trace['z'])
-                self.inst[:,'latitude'] = latitude
-                self.inst[:,'longitude'] = longitude
-                self.inst[:,'altitude'] = altitude
-                # store values for plotting locations for vectors
-                self.inst[:,'x'] = trace['x'].values
-                self.inst[:,'y'] = trace['y'].values
-                self.inst[:,'z'] = trace['z'].values
-                self.inst.data = self.inst[self.inst['altitude'] > 250.]
-
-                # also need to provide transformation from ECEF to S/C
-                # going to leave that a null transformation so we can plot in ECF
-                self.inst[:,'sc_xhat_x'], self.inst[:,'sc_xhat_y'], self.inst[:,'sc_xhat_z'] = 1., 0., 0.
-                self.inst[:,'sc_yhat_x'], self.inst[:,'sc_yhat_y'], self.inst[:,'sc_yhat_z'] = 0., 1., 0.
-                self.inst[:,'sc_zhat_x'], self.inst[:,'sc_zhat_y'], self.inst[:,'sc_zhat_z'] = 0., 0., 1.
-                self.inst.data.index = pysat.utils.time.create_date_range(pysat.datetime(2000,1,1),
-                                                                    pysat.datetime(2000,1,1)+pds.DateOffset(seconds=len(self.inst.data)-1),
-                                                                    freq='S')
-                pymv.satellite.add_mag_drift_unit_vectors(self.inst)
-
-                #if i % 2 == 0:
-                length = 500
-                vx = self.inst['unit_zon_x']
-                vy = self.inst['unit_zon_y']
-                vz = self.inst['unit_zon_z']
-                if not on_travis:
-                    ax.quiver3D(self.inst['x'] + length*vx, self.inst['y'] + length*vy,
-                                self.inst['z'] + length*vz, vx, vy,vz, length=500.,
-                                color='green') #, pivot='tail')
-                length = 500
-                vx = self.inst['unit_fa_x']
-                vy = self.inst['unit_fa_y']
-                vz = self.inst['unit_fa_z']
-                if not on_travis:
-                    ax.quiver3D(self.inst['x'] + length*vx, self.inst['y'] + length*vy,
-                                self.inst['z'] + length*vz, vx, vy,vz, length=500.,
-                                color='purple') #, pivot='tail')
-                length = 500
-                vx = self.inst['unit_mer_x']
-                vy = self.inst['unit_mer_y']
-                vz = self.inst['unit_mer_z']
-                if not on_travis:
-                    ax.quiver3D(self.inst['x'] + length*vx, self.inst['y'] + length*vy,
-                                self.inst['z'] + length*vz, vx, vy,vz, length=500.,
-                                color='red') #, pivot='tail')
-
-                # check that vectors norm to 1
-                assert np.all(np.sqrt(self.inst['unit_zon_x']**2 + self.inst['unit_zon_y']**2 + self.inst['unit_zon_z']**2) > 0.999999)
-                assert np.all(np.sqrt(self.inst['unit_fa_x']**2 + self.inst['unit_fa_y']**2 + self.inst['unit_fa_z']**2) > 0.999999)
-                assert np.all(np.sqrt(self.inst['unit_mer_x']**2 + self.inst['unit_mer_y']**2 + self.inst['unit_mer_z']**2) > 0.999999)
-                # confirm vectors are mutually orthogonal
-                dot1 =  self.inst['unit_zon_x']*self.inst['unit_fa_x'] + self.inst['unit_zon_y']*self.inst['unit_fa_y']  + self.inst['unit_zon_z']*self.inst['unit_fa_z']
-                dot2 =  self.inst['unit_zon_x']*self.inst['unit_mer_x'] + self.inst['unit_zon_y']*self.inst['unit_mer_y']  + self.inst['unit_zon_z']*self.inst['unit_mer_z']
-                dot3 =  self.inst['unit_fa_x']*self.inst['unit_mer_x'] + self.inst['unit_fa_y']*self.inst['unit_mer_y']  + self.inst['unit_fa_z']*self.inst['unit_mer_z']
-                assert np.all(np.abs(dot1) < 1.E-6)
-                assert np.all(np.abs(dot2) < 1.E-6)
-                assert np.all(np.abs(dot3) < 1.E-6)
-
-                # ensure that zonal vector is generally eastward
-                ones = np.ones(len(self.inst.data.index))
-                zeros = np.zeros(len(self.inst.data.index))
-                ex, ey, ez = pymv.enu_to_ecef_vector(ones, zeros, zeros, self.inst['latitude'], self.inst['longitude'])
-                nx, ny, nz = pymv.enu_to_ecef_vector(zeros, ones, zeros, self.inst['latitude'], self.inst['longitude'])
-                ux, uy, uz = pymv.enu_to_ecef_vector(zeros, zeros, ones, self.inst['latitude'], self.inst['longitude'])
-
-                dot1 =  self.inst['unit_zon_x']*ex + self.inst['unit_zon_y']*ey  + self.inst['unit_zon_z']*ez
-                assert np.all(dot1 > 0.)
-
-                dot1 =  self.inst['unit_fa_x']*nx + self.inst['unit_fa_y']*ny  + self.inst['unit_fa_z']*nz
-                assert np.all(dot1 > 0.)
-
-                dot1 =  self.inst['unit_mer_x']*ux + self.inst['unit_mer_y']*uy  + self.inst['unit_mer_z']*uz
-                assert np.all(dot1 > 0.)
-
-
-            if not on_travis:
-                plt.savefig(''.join(('magnetic_unit_vectors_',str(int(p_lat)),'.pdf')))
-                plt.close()
-        ## plot Earth
-        #u = np.linspace(0, 2 * np.pi, 100)
-        #v = np.linspace(60.*np.pi/180., 120.*np.pi/180., 100)
-        #xx = 6371. * np.outer(np.cos(u), np.sin(v))
-        #yy = 6371. * np.outer(np.sin(u), np.sin(v))
-        #zz = 6371. * np.outer(np.ones(np.size(u)), np.cos(v))
-        #ax.plot_surface(xx, yy, zz, rstride=4, cstride=4, color='darkgray')
-        #plt.savefig('magnetic_unit_vectors_w_globe.pdf')
-        #print truthiness
-        assert np.all(truthiness)
 
 #     def test_geomag_drift_scalars_accuracy_vs_settings(self):
 #         import matplotlib.pyplot as plt
@@ -911,651 +277,8 @@ class TestCore():
 #         n_z_diff = north_zonal[:,:,1:] - north_zonal[:,:,:-1]
 #         n_z_diff.mean(axis=0).mean(axis=0)
 #
-    def test_apex_plots(self):
-        import matplotlib.pyplot as plt
-        import os
-        # on_travis = os.environ.get('ONTRAVIS') == 'True'
 
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(120.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-        # set the date
-        date = datetime.datetime(2000,1,1)
-        # memory for results
-        apex_lat = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_lon = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_alt = np.zeros((len(p_lats), len(p_longs)+1))
 
-
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                # iterate through target cyclicly and run commands
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.apex_location_info, [p_lat]*len(p_longs), p_longs,
-                                                                            p_alts, [date]*len(p_longs),
-                                                                            return_geodetic=True))
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output
-                x, y, z, olat, olon, oalt = pending.pop(0).get()
-                apex_lat[i,:-1] = olat
-                apex_lon[i,:-1] = olon
-                apex_alt[i,:-1] = oalt
-
-        else:
-            # single processor case
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                x, y, z, olat, olon, oalt = pymv.apex_location_info([p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        return_geodetic=True)
-                apex_lat[i,:-1] = olat
-                apex_lon[i,:-1] = olon
-                apex_alt[i,:-1] = oalt
-
-        # calculate difference between apex longitude and original longitude
-        # values for apex long are -180 to 180, shift to 0 to 360
-        # process degrees a bit to make the degree difference the most meaningful (close to 0)
-        idx, idy, = np.where(apex_lon < 0.)
-        apex_lon[idx, idy] += 360.
-        idx, idy, = np.where(apex_lon >= 360.)
-        apex_lon[idx, idy] -= 360.
-        apex_lon[:, :-1] -= p_longs
-        idx, idy, = np.where(apex_lon > 180.)
-        apex_lon[idx, idy] -= 360.
-        idx, idy, = np.where(apex_lon <= -180.)
-        apex_lon[idx, idy] += 360.
-
-        # account for periodicity
-        apex_lat[:,-1] = apex_lat[:,0]
-        apex_lon[:,-1] = apex_lon[:,0]
-        apex_alt[:,-1] = apex_alt[:,0]
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-        ytickvals = ['-25', '-12.5', '0', '12.5', '25']
-
-        try:
-            fig = plt.figure()
-            plt.imshow(apex_lat, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Apex Latitude (Degrees) at 120 km')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_lat.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(apex_lon, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Apex Longitude Difference (Degrees) at 120 km')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_lon.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_alt), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Altitude (km) at 120 km')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_alt.pdf')
-            plt.close()
-        except:
-            pass
-
-    def test_apex_diff_plots(self):
-        import matplotlib.pyplot as plt
-        import os
-        # on_travis = os.environ.get('ONTRAVIS') == 'True'
-
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-        # set the date
-        date = datetime.datetime(2000,1,1)
-        # memory for results
-        apex_lat = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_lon = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_alt = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_z = np.zeros((len(p_lats), len(p_longs)+1))
-        norm_alt = np.zeros((len(p_lats), len(p_longs)+1))
-
-
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                # iterate through target cyclicly and run commands
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.apex_location_info, [p_lat]*len(p_longs), p_longs,
-                                                                            p_alts, [date]*len(p_longs),
-                                                                            fine_step_size=1.E-5,
-                                                                            return_geodetic=True))
-                pending.append(dview.apply_async(pymv.apex_location_info, [p_lat]*len(p_longs), p_longs,
-                                                                            p_alts, [date]*len(p_longs),
-                                                                            fine_step_size=5.E-6,
-                                                                            return_geodetic=True))
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output
-                x, y, z, _, _, h = pending.pop(0).get()
-                x2, y2, z2, _, _, h2 = pending.pop(0).get()
-                apex_lat[i,:-1] = np.abs(x2 - x)
-                apex_lon[i,:-1] = np.abs(y2 - y)
-                apex_z[i,:-1] = np.abs(z2 - z)
-                apex_alt[i,:-1] = np.abs(h2 - h)
-
-        else:
-            # single processor case
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                x, y, z, _, _, h = pymv.apex_location_info([p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        fine_step_size=1.E-5, return_geodetic=True)
-                x2, y2, z2, _, _, h2 = pymv.apex_location_info([p_lat]*len(p_longs), p_longs,
-                                                                           p_alts, [date]*len(p_longs),
-                                                                           fine_step_size=5.E-6, return_geodetic=True)
-
-                norm_alt[i,:-1] = h
-                apex_lat[i,:-1] = np.abs(x2 - x)
-                apex_lon[i,:-1] = np.abs(y2 - y)
-                apex_z[i,:-1] = np.abs(z2 - z)
-                apex_alt[i,:-1] = np.abs(h2 - h)
-
-        # account for periodicity
-        apex_lat[:,-1] = apex_lat[:,0]
-        apex_lon[:,-1] = apex_lon[:,0]
-        apex_z[:,-1] = apex_z[:,0]
-        apex_alt[:,-1] = apex_alt[:,0]
-        norm_alt[:,-1] = norm_alt[:,0]
-
-        idx, idy, = np.where(apex_lat > 10.)
-        print('Locations with large apex x (ECEF) location differences.', p_lats[idx], p_longs[idx])
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-        ytickvals = ['-50', '-25', '0', '25', '50']
-
-        try:
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_lat), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Location Difference (ECEF-x km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_loc_diff_x.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_lon), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Location Difference (ECEF-y km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_loc_diff_y.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_z), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Location Difference (ECEF-z km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_loc_diff_z.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_alt/norm_alt), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Altitude Normalized Difference (km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_norm_loc_diff_h.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_alt), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Altitude Difference (km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_loc_diff_h.pdf')
-            plt.close()
-
-        except:
-            pass
-
-### Test apex location info for sensitivity to fine_steps parameters
-    def test_apex_fine_max_step_diff_plots(self):
-        import matplotlib.pyplot as plt
-        import os
-        # on_travis = os.environ.get('ONTRAVIS') == 'True'
-
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-        # set the date
-        date = datetime.datetime(2000,1,1)
-        # memory for results
-        apex_lat = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_lon = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_alt = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_z = np.zeros((len(p_lats), len(p_longs)+1))
-        norm_alt = np.zeros((len(p_lats), len(p_longs)+1))
-
-
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                # iterate through target cyclicly and run commands
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.apex_location_info, [p_lat]*len(p_longs), p_longs,
-                                                                            p_alts, [date]*len(p_longs),
-                                                                            fine_max_steps=5,
-                                                                            return_geodetic=True))
-                pending.append(dview.apply_async(pymv.apex_location_info, [p_lat]*len(p_longs), p_longs,
-                                                                            p_alts, [date]*len(p_longs),
-                                                                            fine_max_steps=10,
-                                                                            return_geodetic=True))
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output
-                x, y, z, _, _, h = pending.pop(0).get()
-                x2, y2, z2, _, _, h2 = pending.pop(0).get()
-                apex_lat[i,:-1] = np.abs(x2 - x)
-                apex_lon[i,:-1] = np.abs(y2 - y)
-                apex_z[i,:-1] = np.abs(z2 - z)
-                apex_alt[i,:-1] = np.abs(h2 - h)
-
-        else:
-            # single processor case
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                x, y, z, _, _, h = pymv.apex_location_info([p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        fine_max_steps=5, return_geodetic=True)
-                x2, y2, z2, _, _, h2 = pymv.apex_location_info([p_lat]*len(p_longs), p_longs,
-                                                                           p_alts, [date]*len(p_longs),
-                                                                           fine_max_steps=10, return_geodetic=True)
-
-                norm_alt[i,:-1] = h
-                apex_lat[i,:-1] = np.abs(x2 - x)
-                apex_lon[i,:-1] = np.abs(y2 - y)
-                apex_z[i,:-1] = np.abs(z2 - z)
-                apex_alt[i,:-1] = np.abs(h2 - h)
-
-        # account for periodicity
-        apex_lat[:,-1] = apex_lat[:,0]
-        apex_lon[:,-1] = apex_lon[:,0]
-        apex_z[:,-1] = apex_z[:,0]
-        apex_alt[:,-1] = apex_alt[:,0]
-        norm_alt[:,-1] = norm_alt[:,0]
-
-        idx, idy, = np.where(apex_lat > 10.)
-        print('Locations with large apex x (ECEF) location differences.', p_lats[idx], p_longs[idx])
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-        ytickvals = ['-50', '-25', '0', '25', '50']
-
-        try:
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_lat), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Location Difference (ECEF-x km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_loc_max_steps_diff_x.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_lon), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Location Difference (ECEF-y km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_loc_max_steps_diff_y.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_z), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Location Difference (ECEF-z km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_loc_max_steps_diff_z.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_alt/norm_alt), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Altitude Normalized Difference (km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_norm_loc_max_steps_diff_h.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_alt), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Apex Altitude Normalized Difference (km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('apex_loc_max_steps_diff_h.pdf')
-            plt.close()
-
-        except:
-            pass
-
-    def test_ecef_geodetic_diff_plots(self):
-        import matplotlib.pyplot as plt
-        import os
-        # on_travis = os.environ.get('ONTRAVIS') == 'True'
-
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-        # set the date
-        date = datetime.datetime(2000,1,1)
-        # memory for results
-        apex_x = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_y = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_z = np.zeros((len(p_lats), len(p_longs)+1))
-        norm_alt = np.zeros((len(p_lats), len(p_longs)+1))
-
-
-#         # set up multi
-#         if self.dc is not None:
-#             import itertools
-#             targets = itertools.cycle(dc.ids)
-#             pending = []
-#             for i,p_lat in enumerate(p_lats):
-#                 print (i, p_lat)
-#                 # iterate through target cyclicly and run commands
-#                 dview.targets = targets.next()
-#                 pending.append(dview.apply_async(pymv.geodetic_to_ecef, [p_lat]*len(p_longs), p_longs,
-#                                                                             p_alts))
-#             for i,p_lat in enumerate(p_lats):
-#                 print ('collecting ', i, p_lat)
-#                 # collect output
-#                 x, y, z = pending.pop(0).get()
-#
-#                 # iterate through target cyclicly and run commands
-#                 dview.targets = targets.next()
-#                 pending.append(dview.apply_async(pymv.ecef_to_geodetic, x, y, z))
-#
-#             for i,p_lat in enumerate(p_lats):
-#                 print ('collecting 2', i, p_lat)
-#                 # collect output
-#                 lat2, lon2, alt2 = pending.pop(0).get()
-#
-#                 # iterate through target cyclicly and run commands
-#                 dview.targets = targets.next()
-#                 pending.append(dview.apply_async(pymv.geodetic_to_ecef, lat2, lon2, alt2))
-#
-#             for i,p_lat in enumerate(p_lats):
-#                 print ('collecting 3', i, p_lat)
-#                 x2, y2, z2, _, _, h2 = pending.pop(0).get()
-#                 apex_x[i,:-1] = np.abs(x2 - x)
-#                 apex_y[i,:-1] = np.abs(y2 - y)
-#                 apex_z[i,:-1] = np.abs(z2 - z)
-#
-#
-#         else:
-        # single processor case
-        for i,p_lat in enumerate(p_lats):
-            print (i, p_lat)
-            x, y, z = pymv.geodetic_to_ecef([p_lat]*len(p_longs), p_longs, p_alts)
-            lat2, lon2, alt2 = pymv.ecef_to_geodetic(x, y, z)
-            x2, y2, z2 = pymv.geodetic_to_ecef(lat2, lon2, alt2)
-            apex_x[i,:-1] = np.abs(x2 - x)
-            apex_y[i,:-1] = np.abs(y2 - y)
-            apex_z[i,:-1] = np.abs(z2 - z)
-
-
-        # account for periodicity
-        apex_x[:,-1] = apex_x[:,0]
-        apex_y[:,-1] = apex_y[:,0]
-        apex_z[:,-1] = apex_z[:,0]
-
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-        ytickvals = ['-50', '-25', '0', '25', '50']
-
-        try:
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_x), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log ECEF-Geodetic Location Difference (ECEF-x km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('ecef_geodetic_diff_x.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_y), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log ECEF-Geodetic Location Difference (ECEF-y km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('ecef_geodetic_diff_y.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_z), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log ECEF-Geodetic Location Difference (ECEF-z km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('ecef_geodetic_diff_z.pdf')
-            plt.close()
-
-        except:
-            pass
-
-    def test_ecef_geodetic_apex_diff_plots(self):
-        import matplotlib.pyplot as plt
-        import os
-        # on_travis = os.environ.get('ONTRAVIS') == 'True'
-
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-        # set the date
-        date = datetime.datetime(2000,1,1)
-        # memory for results
-        apex_x = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_y = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_z = np.zeros((len(p_lats), len(p_longs)+1))
-        apex_alt = np.zeros((len(p_lats), len(p_longs)+1))
-        norm_alt = np.zeros((len(p_lats), len(p_longs)+1))
-
-
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                # iterate through target cyclicly and run commands
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.geodetic_to_ecef, np.array([p_lat]*len(p_longs)), p_longs,
-                                                                            p_alts))
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output
-                x, y, z = pending.pop(0).get()
-
-                # iterate through target cyclicly and run commands
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.python_ecef_to_geodetic, x, y, z))
-
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting 2', i, p_lat)
-                # collect output
-                lat2, lon2, alt2 = pending.pop(0).get()
-
-                # iterate through target cyclicly and run commands
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.apex_location_info, np.array([p_lat]*len(p_longs)), p_longs,
-                                                                            p_alts, [date]*len(p_longs),
-                                                                            return_geodetic=True))
-
-                pending.append(dview.apply_async(pymv.apex_location_info, lat2, lon2, alt2,
-                                                                            [date]*len(p_longs),
-                                                                            return_geodetic=True))
-
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting 3', i, p_lat)
-                x, y, z, _, _, h = pending.pop(0).get()
-                x2, y2, z2, _, _, h2 = pending.pop(0).get()
-                norm_alt[i, :-1] = np.abs(h)
-                apex_x[i,:-1] = np.abs(x2 - x)
-                apex_y[i,:-1] = np.abs(y2 - y)
-                apex_z[i,:-1] = np.abs(z2 - z)
-                apex_alt[i, :-1] = np.abs(h2 - h)
-
-
-        else:
-            # single processor case
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                x, y, z = pymv.geodetic_to_ecef([p_lat]*len(p_longs), p_longs, p_alts)
-                lat2, lon2, alt2 = pymv.ecef_to_geodetic(x, y, z)
-                x2, y2, z2 = pymv.geodetic_to_ecef(lat2, lon2, alt2)
-                apex_x[i,:-1] = np.abs(x2 - x)
-                apex_y[i,:-1] = np.abs(y2 - y)
-                apex_z[i,:-1] = np.abs(z2 - z)
-
-
-        # account for periodicity
-        apex_x[:,-1] = apex_x[:,0]
-        apex_y[:,-1] = apex_y[:,0]
-        apex_z[:,-1] = apex_z[:,0]
-        apex_alt[:,-1] = apex_alt[:,0]
-        norm_alt[:,-1] = norm_alt[:,0]
-
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-        ytickvals = ['-50', '-25', '0', '25', '50']
-
-        try:
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_x), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log ECEF-Geodetic Apex Difference (ECEF-x km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('ecef_geodetic_apex_diff_x.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_y), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log ECEF-Geodetic Apex Difference (ECEF-y km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('ecef_geodetic_apex_diff_y.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_z), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log ECEF-Geodetic Apex Difference (ECEF-z km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('ecef_geodetic_apex_diff_z.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_alt), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log ECEF-Geodetic Apex Altitude Difference (km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('ecef_geodetic_apex_diff_h.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(apex_alt/norm_alt), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ytickvals)
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log ECEF-Geodetic Apex Normalized Altitude Difference (km)')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('ecef_geodetic_apex_norm_diff_h.pdf')
-            plt.close()
-
-        except:
-            pass
 
     def test_aaaaaaa_unit_vector_component_plots(self):
         import matplotlib.pyplot as plt
@@ -1607,6 +330,7 @@ class TestCore():
                 zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
                 bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
                 mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
+
                 # pull out info about the vector generation
                 grad_zon[i,:-1], grad_mer[i,:-1] = infod['diff_zonal_apex'], infod['diff_mer_apex']
                 tol_zon[i,:-1], tol_mer[i,:-1] = infod['diff_zonal_vec'], infod['diff_mer_vec']
@@ -2286,435 +1010,7 @@ class TestCore():
         assert np.all(np.abs(tol_mer) <= 1.E-4)
         assert np.all(np.abs(grad_zon) <= 1.E-4)
 
-    def test_integrated_unit_vector_component_plots(self):
-        import matplotlib.pyplot as plt
 
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-        zvx = np.zeros((len(p_lats), len(p_longs)+1))
-        zvy = zvx.copy(); zvz = zvx.copy()
-        mx = zvx.copy(); my = zvx.copy(); mz = zvx.copy()
-        bx = zvx.copy(); by = zvx.copy(); bz = zvx.copy()
-        date = datetime.datetime(2000,1,1)
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                # iterate through target cyclicly and run commands
-                print (i, p_lat)
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.calculate_integrated_mag_drift_unit_vectors_ecef,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        steps=None, max_steps=10000, step_size=10.,
-                                                                        ref_height=120.))
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                    # collect output
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pending.pop(0).get()
-                zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-        else:
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pymv.calculate_integrated_mag_drift_unit_vectors_ecef([p_lat]*len(p_longs), p_longs,
-                                                                                        p_alts, [date]*len(p_longs),
-                                                                                        steps=None, max_steps=10000, step_size=10.,
-                                                                                        ref_height=120.)
-                zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-
-        # account for periodicity
-        zvx[:,-1] = zvx[:,0]
-        zvy[:,-1] = zvy[:,0]
-        zvz[:,-1] = zvz[:,0]
-        bx[:,-1] = bx[:,0]
-        by[:,-1] = by[:,0]
-        bz[:,-1] = bz[:,0]
-        mx[:,-1] = mx[:,0]
-        my[:,-1] = my[:,0]
-        mz[:,-1] = mz[:,0]
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-
-        try:
-            fig = plt.figure()
-            plt.imshow(zvx, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Zonal Unit Vector - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_east.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(zvy, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Zonal Unit Vector - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_north.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(zvz, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Zonal Unit Vector - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_up.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(bx, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Field Aligned Unit Vector - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_east.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(by, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Field Aligned Unit Vector - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_north.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(bz, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Field Aligned Unit Vector - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_up.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(mx, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Meridional Unit Vector - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_east.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(my, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Meridional Unit Vector - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_north.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(mz, origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Meridional Unit Vector - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_up.pdf')
-            plt.close()
-        except:
-            pass
-
-    def closed_loop_footpoint_sensitivity_plots(self, direction, vector_direction):
-        import matplotlib.pyplot as plt
-        # from mpl_toolkits.mplot3d import Axes3D
-        import os
-
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-
-        # zonal vector components
-        # +1 on length of longitude array supports repeating first element
-        # shows nice periodicity on the plots
-        zvx = np.zeros((len(p_lats), len(p_longs)+1))
-        #  meridional vecrtor components
-        mx = zvx.copy(); my = zvx.copy(); mz = zvx.copy()
-        date = datetime.datetime(2000,1,1)
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                # iterate through target cyclicly and run commands
-                print (i, p_lat)
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.closed_loop_edge_lengths_via_footpoint,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        direction,
-                                                                        vector_direction,
-                                                                        edge_length=25.,
-                                                                        edge_steps=5))
-                pending.append(dview.apply_async(pymv.closed_loop_edge_lengths_via_footpoint,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        direction,
-                                                                        vector_direction,
-                                                                        edge_length=25.,
-                                                                        edge_steps=10))
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output from first run
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pending.pop(0).get()
-                # collect output from second run
-                _a, _b, _c = pending.pop(0).get()
-                # take difference with first run
-                mx[i,:-1] = (mx[i,:-1] - _a)/mx[i,:-1]
-                my[i,:-1] = (my[i,:-1] - _b)/my[i,:-1]
-                mz[i,:-1] = (mz[i,:-1] - _c)/mz[i,:-1]
-
-        else:
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.closed_loop_edge_lengths_via_footpoint([p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        direction,
-                                                                        vector_direction,
-                                                                        edge_length=25.,
-                                                                        edge_steps=5)
-
-                # second run
-                _a, _b, _c = pymv.closed_loop_edge_lengths_via_footpoint([p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        direction,
-                                                                        vector_direction,
-                                                                        edge_length=25.,
-                                                                        edge_steps=10)
-                # take difference with first run
-                mx[i,:-1] = (mx[i,:-1] - _a)/mx[i,:-1]
-                my[i,:-1] = (my[i,:-1] - _b)/my[i,:-1]
-                mz[i,:-1] = (mz[i,:-1] - _c)/mz[i,:-1]
-
-
-        # account for periodicity
-        mx[:,-1] = mx[:,0]
-        my[:,-1] = my[:,0]
-        mz[:,-1] = mz[:,0]
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-
-        try:
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Closed Loop Edge Length Normalized Difference, Footpoint Path')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig(direction+'_'+vector_direction+'_closed_loop_footpoint_edge_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(my)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Closed Loop Closest Approach Normalized Difference, Pos Footpoint Path')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig(direction+'_'+vector_direction+'_closed_loop_footpoint_pos_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Closed Loop Closest Approach Normalized Difference, Minus Footpoint Path')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig(direction+'_'+vector_direction+'_closed_loop_footpoint_min_diff.pdf')
-            plt.close()
-
-            # calculate mean and standard deviation and then plot those
-            plt.figure()
-            yerrx = np.nanstd(np.log10(mx[:,:-1]), axis=0)
-            yerry = np.nanstd(np.log10(my[:,:-1]), axis=0)
-            yerrz = np.nanstd(np.log10(mz[:,:-1]), axis=0)
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mx[:,:-1]), axis=0)),
-                            yerr=yerrx, label='Edge')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(my[:,:-1]), axis=0)),
-                            yerr=yerry, label='Positive')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mz[:,:-1]), axis=0)),
-                            yerr=yerrz, label='Minus')
-            plt.xlabel('Longitude (Degrees)')
-            plt.ylabel('Log Normalized Closed Loop Difference')
-            plt.title("Sensitivity of Closed Loop Values")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig(direction+'_'+vector_direction+'_length_diff_v_longitude.pdf' )
-            plt.close()
-
-        except:
-            pass
-
-    # def test_closed_loop_footpoint_sensitivity_plots(self):
-    #     f = functools.partial(self.closed_loop_footpoint_sensitivity_plots, 'north', 'meridional')
-    #     yield(f,)
-    #     f = functools.partial(self.closed_loop_footpoint_sensitivity_plots, 'south', 'meridional')
-    #     yield(f,)
-    #     f = functools.partial(self.closed_loop_footpoint_sensitivity_plots, 'north', 'zonal')
-    #     yield(f,)
-    #     f = functools.partial(self.closed_loop_footpoint_sensitivity_plots, 'south', 'zonal')
-    #     yield(f,)
-
-
-    def closed_loop_footpoint_value_plots(self, direction, vector_direction):
-        import matplotlib.pyplot as plt
-        # from mpl_toolkits.mplot3d import Axes3D
-        import os
-
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-
-        # zonal vector components
-        # +1 on length of longitude array supports repeating first element
-        # shows nice periodicity on the plots
-        zvx = np.zeros((len(p_lats), len(p_longs)+1))
-        #  meridional vecrtor components
-        mx = zvx.copy(); my = zvx.copy(); mz = zvx.copy()
-        date = datetime.datetime(2000,1,1)
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                # iterate through target cyclicly and run commands
-                print (i, p_lat)
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.closed_loop_edge_lengths_via_footpoint,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        direction,
-                                                                        vector_direction,
-                                                                        edge_length=25.,
-                                                                        edge_steps=5))
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output from first run
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pending.pop(0).get()
-
-        else:
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.closed_loop_edge_lengths_via_footpoint([p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        direction,
-                                                                        vector_direction,
-                                                                        edge_length=25.,
-                                                                        edge_steps=5)
-
-        # account for periodicity
-        mx[:,-1] = mx[:,0]
-        my[:,-1] = my[:,0]
-        mz[:,-1] = mz[:,0]
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-
-        try:
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Closed Loop Edge Length, Footpoint Path')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig(direction+'_'+vector_direction+'_closed_loop_footpoint_edge.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(my)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Closed Loop Closest Approach, Pos Footpoint Path')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig(direction+'_'+vector_direction+'_closed_loop_footpoint_pos.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Closed Loop Closest Approach, Minus Footpoint Path')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig(direction+'_'+vector_direction+'_closed_loop_footpoint_min.pdf')
-            plt.close()
-
-            # calculate mean and standard deviation and then plot those
-            plt.figure()
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mx[:,:-1]), axis=0)),
-                            yerr=np.abs(np.log10(mx.std(axis=0)) - np.log10(mx.median(axis=0))), label='Edge')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(my[:,:-1]), axis=0)),
-                            yerr=np.abs(np.log10(my.std(axis=0)) - np.log10(my.median(axis=0))), label='Positive')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(my[:,:-1]), axis=0)),
-                            yerr=np.abs(np.log10(mz.std(axis=0)) - np.log10(mz.median(axis=0))), label='Minus')
-            plt.xlabel('Longitude (Degrees)')
-            plt.ylabel('Log Closed Loop Values')
-            plt.title("Closed Loop Values")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig(direction+'_'+vector_direction+'_length_v_longitude.pdf' )
-            plt.close()
-
-        except:
-            pass
-
-
-    # def test_closed_loop_footpoint_value_plots(self):
-    #     f = functools.partial(self.closed_loop_footpoint_value_plots, 'north', 'meridional')
-    #     yield(f,)
-    #     f = functools.partial(self.closed_loop_footpoint_value_plots, 'south', 'meridional')
-    #     yield(f,)
-    #     f = functools.partial(self.closed_loop_footpoint_value_plots, 'north', 'zonal')
-    #     yield(f,)
-    #     f = functools.partial(self.closed_loop_footpoint_value_plots, 'south', 'zonal')
-    #     yield(f,)
 
 
     def test_unit_vector_component_stepsize_sensitivity_plots(self):
@@ -2972,818 +1268,9 @@ class TestCore():
         except:
             print('Skipping plots due to error.')
 
-    def test_integrated_unit_vector_component_stepsize_sensitivity_plots(self):
-        import matplotlib.pyplot as plt
-        # from mpl_toolkits.mplot3d import Axes3D
-        import os
 
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
 
-        # zonal vector components
-        # +1 on length of longitude array supports repeating first element
-        # shows nice periodicity on the plots
-        zvx = np.zeros((len(p_lats), len(p_longs)+1))
-        zvy = zvx.copy(); zvz = zvx.copy()
-        #  meridional vecrtor components
-        mx = zvx.copy(); my = zvx.copy(); mz = zvx.copy()
-        # field aligned, along B
-        bx = zvx.copy(); by = zvx.copy(); bz = zvx.copy()
-        date = datetime.datetime(2000,1,1)
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                # iterate through target cyclicly and run commands
-                print (i, p_lat)
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.calculate_integrated_mag_drift_unit_vectors_ecef,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        steps=None, max_steps=10000, step_size=10.,
-                                                                        ref_height=120.))
-                pending.append(dview.apply_async(pymv.calculate_integrated_mag_drift_unit_vectors_ecef,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        steps=None, max_steps=1000, step_size=100.,
-                                                                        ref_height=120.))
 
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output from first run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pending.pop(0).get()
-                zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # collect output from second run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pending.pop(0).get()
-                _a, _b, _c = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                zvx[i,:-1] = (zvx[i,:-1] - _a)/zvx[i,:-1]
-                zvy[i,:-1] = (zvy[i,:-1] - _b)/zvy[i,:-1]
-                zvz[i,:-1] = (zvz[i,:-1] - _c)/zvz[i,:-1]
-
-                _a, _b, _c  = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                bx[i,:-1] = (bx[i,:-1] - _a)/bx[i,:-1]
-                by[i,:-1] = (by[i,:-1] - _b)/by[i,:-1]
-                bz[i,:-1] = (bz[i,:-1] - _c)/bz[i,:-1]
-
-                _a, _b, _c = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                mx[i,:-1] = (mx[i,:-1] - _a)/mx[i,:-1]
-                my[i,:-1] = (my[i,:-1] - _b)/my[i,:-1]
-                mz[i,:-1] = (mz[i,:-1] - _c)/mz[i,:-1]
-
-        else:
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pymv.calculate_integrated_mag_drift_unit_vectors_ecef([p_lat]*len(p_longs), p_longs,
-                                                                                        p_alts, [date]*len(p_longs),
-                                                                                        steps=None, max_steps=10000, step_size=10.,
-                                                                                        ref_height=120.)
-                zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-
-                # second run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pymv.calculate_integrated_mag_drift_unit_vectors_ecef([p_lat]*len(p_longs), p_longs,
-                                                                                        p_alts, [date]*len(p_longs),
-                                                                                        steps=None, max_steps=1000, step_size=100.,
-                                                                                        ref_height=120.)
-                _a, _b, _c = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                zvx[i,:-1] = (zvx[i,:-1] - _a)/zvx[i,:-1]
-                zvy[i,:-1] = (zvy[i,:-1] - _b)/zvy[i,:-1]
-                zvz[i,:-1] = (zvz[i,:-1] - _c)/zvz[i,:-1]
-
-                _a, _b, _c  = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                bx[i,:-1] = (bx[i,:-1] - _a)/bx[i,:-1]
-                by[i,:-1] = (by[i,:-1] - _b)/by[i,:-1]
-                bz[i,:-1] = (bz[i,:-1] - _c)/bz[i,:-1]
-
-                _a, _b, _c = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                mx[i,:-1] = (mx[i,:-1] - _a)/mx[i,:-1]
-                my[i,:-1] = (my[i,:-1] - _b)/my[i,:-1]
-                mz[i,:-1] = (mz[i,:-1] - _c)/mz[i,:-1]
-
-
-        # account for periodicity
-        zvx[:,-1] = zvx[:,0]
-        zvy[:,-1] = zvy[:,0]
-        zvz[:,-1] = zvz[:,0]
-        bx[:,-1] = bx[:,0]
-        by[:,-1] = by[:,0]
-        bz[:,-1] = bz[:,0]
-        mx[:,-1] = mx[:,0]
-        my[:,-1] = my[:,0]
-        mz[:,-1] = mz[:,0]
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-
-        try:
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Nornalized Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_east_norm_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvy)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Nornalized Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_north_norm_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Nornalized Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_up_norm_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(bx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Nornalized Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_east_norm_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(by)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Nornalized Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_north_norm_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(bz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Nornalized Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_up_norm_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Nornalized Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_east_norm_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(my)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Nornalized Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_north_norm_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Nornalized Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_up_norm_diff.pdf')
-            plt.close()
-
-            # calculate mean and standard deviation and then plot those
-            # print (p_longs)
-            # print (np.nanmean(np.abs(zvx), axis=0))
-            # print (np.nanstd(np.abs(zvx), axis=0))
-            plt.figure()
-            yerrx = np.nanstd(np.log10(zvx[:,:-1]), axis=0)
-            yerry = np.nanstd(np.log10(zvy[:,:-1]), axis=0)
-            yerrz = np.nanstd(np.log10(zvz[:,:-1]), axis=0)
-
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvx[:,:-1]), axis=0)),
-                            yerr=yerrx, label='East')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvy[:,:-1]), axis=0)),
-                            yerr=yerry, label='North')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvz[:,:-1]), axis=0)),
-                            yerr=yerrz, label='Up')
-            plt.xlabel('Longitude (Degrees)')
-            plt.ylabel('Log Normalized Change in Zonal Vector')
-            plt.title("Sensitivity of Zonal Unit Vector")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('int_zonal_diff_v_longitude.pdf' )
-            plt.close()
-
-            # # calculate mean and standard deviation and then plot those
-            # plt.figure()
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(bx[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(bx[:,:-1])), axis=0), label='East')
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(by[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(by[:,:-1])), axis=0), label='North')
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(bz[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(bz[:,:-1])), axis=0), label='Up')
-            # plt.xlabel('Longitude (Degrees)')
-            # plt.ylabel('Log Normalized Change in Field-Aligned Vector')
-            # plt.title("Sensitivity of Field-Aligned Unit Vector")
-            # plt.legend()
-            # plt.tight_layout()
-            # plt.savefig('int_fa_diff_v_longitude.pdf' )
-            # plt.close()
-
-            # calculate mean and standard deviation and then plot those
-            plt.figure()
-            yerrx = np.nanstd(np.log10(mx[:,:-1]), axis=0)
-            yerry = np.nanstd(np.log10(my[:,:-1]), axis=0)
-            yerrz = np.nanstd(np.log10(mz[:,:-1]), axis=0)
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mx[:,:-1]), axis=0)),
-                            yerr=yerrx, label='East')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(my[:,:-1]), axis=0)),
-                            yerr=yerry, label='North')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mz[:,:-1]), axis=0)),
-                            yerr=yerrz, label='Up')
-            plt.xlabel('Longitude (Degrees)')
-            plt.ylabel('Log Normalized Change in Meridional Vector')
-            plt.title("Sensitivity of Meridional Unit Vector")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('int_mer_diff_v_longitude.pdf' )
-            plt.close()
-
-        except:
-            pass
-
-    def test_vector_method_sensitivity_plots(self):
-        import matplotlib.pyplot as plt
-
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-
-        # zonal vector components
-        # +1 on length of longitude array supports repeating first element
-        # shows nice periodicity on the plots
-        zvx = np.zeros((len(p_lats), len(p_longs)+1))
-        zvy = zvx.copy(); zvz = zvx.copy()
-        #  meridional vecrtor components
-        mx = zvx.copy(); my = zvx.copy(); mz = zvx.copy()
-        # field aligned, along B
-        bx = zvx.copy(); by = zvx.copy(); bz = zvx.copy()
-        date = datetime.datetime(2000,1,1)
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                # iterate through target cyclicly and run commands
-                print (i, p_lat)
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.calculate_integrated_mag_drift_unit_vectors_ecef,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs)))
-                pending.append(dview.apply_async(pymv.calculate_mag_drift_unit_vectors_ecef,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs)))
-
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output from first run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pending.pop(0).get()
-                zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # collect output from second run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pending.pop(0).get()
-                _a, _b, _c = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                zvx[i,:-1] = (zvx[i,:-1] - _a)#/zvx[i,:-1]
-                zvy[i,:-1] = (zvy[i,:-1] - _b)#/zvy[i,:-1]
-                zvz[i,:-1] = (zvz[i,:-1] - _c)#/zvz[i,:-1]
-
-                _a, _b, _c  = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                bx[i,:-1] = (bx[i,:-1] - _a)#/bx[i,:-1]
-                by[i,:-1] = (by[i,:-1] - _b)#/by[i,:-1]
-                bz[i,:-1] = (bz[i,:-1] - _c)#/bz[i,:-1]
-
-                _a, _b, _c = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                mx[i,:-1] = (mx[i,:-1] - _a)#/mx[i,:-1]
-                my[i,:-1] = (my[i,:-1] - _b)#/my[i,:-1]
-                mz[i,:-1] = (mz[i,:-1] - _c)#/mz[i,:-1]
-
-        else:
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pymv.calculate_integrated_mag_drift_unit_vectors_ecef([p_lat]*len(p_longs), p_longs,
-                                                                                        p_alts, [date]*len(p_longs))
-                zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-
-                # second run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pymv.calculate_mag_drift_unit_vectors_ecef([p_lat]*len(p_longs), p_longs,
-                                                                                        p_alts, [date]*len(p_longs))
-                _a, _b, _c = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                zvx[i,:-1] = (zvx[i,:-1] - _a)#/zvx[i,:-1]
-                zvy[i,:-1] = (zvy[i,:-1] - _b)#/zvy[i,:-1]
-                zvz[i,:-1] = (zvz[i,:-1] - _c)#/zvz[i,:-1]
-
-                _a, _b, _c  = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                bx[i,:-1] = (bx[i,:-1] - _a)#/bx[i,:-1]
-                by[i,:-1] = (by[i,:-1] - _b)#/by[i,:-1]
-                bz[i,:-1] = (bz[i,:-1] - _c)#/bz[i,:-1]
-
-                _a, _b, _c = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                mx[i,:-1] = (mx[i,:-1] - _a)#/mx[i,:-1]
-                my[i,:-1] = (my[i,:-1] - _b)#/my[i,:-1]
-                mz[i,:-1] = (mz[i,:-1] - _c)#/mz[i,:-1]
-
-
-        # account for periodicity
-        zvx[:,-1] = zvx[:,0]
-        zvy[:,-1] = zvy[:,0]
-        zvz[:,-1] = zvz[:,0]
-        bx[:,-1] = bx[:,0]
-        by[:,-1] = by[:,0]
-        bz[:,-1] = bz[:,0]
-        mx[:,-1] = mx[:,0]
-        my[:,-1] = my[:,0]
-        mz[:,-1] = mz[:,0]
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-
-        try:
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_zonal_east_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvy)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_zonal_north_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_zonal_up_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(bx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_fa_east_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(by)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_fa_north_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(bz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_fa_up_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_mer_east_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(my)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_mer_north_diff.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('method_mer_up_diff.pdf')
-            plt.close()
-
-            # calculate mean and standard deviation and then plot those
-            # print (p_longs)
-            # print (np.nanmean(np.abs(zvx), axis=0))
-            # print (np.nanstd(np.abs(zvx), axis=0))
-            plt.figure()
-            yerrx = np.nanstd(np.log10(zvx[:,:-1]), axis=0)
-            yerry = np.nanstd(np.log10(zvy[:,:-1]), axis=0)
-            yerrz = np.nanstd(np.log10(zvz[:,:-1]), axis=0)
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvx[:,:-1]), axis=0)),
-                            yerr=yerrx, label='East')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvy[:,:-1]), axis=0)),
-                            yerr=yerry, label='North')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvz[:,:-1]), axis=0)),
-                            yerr=yerrz, label='Up')
-            plt.xlabel('Longitude (Degrees)')
-            plt.ylabel('Log Change in Zonal Vector')
-            plt.title("Sensitivity of Zonal Unit Vector")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('method_zonal_diff_v_longitude.pdf' )
-            plt.close()
-
-            # # calculate mean and standard deviation and then plot those
-            # plt.figure()
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(bx[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(bx[:,:-1])), axis=0), label='East')
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(by[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(by[:,:-1])), axis=0), label='North')
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(bz[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(bz[:,:-1])), axis=0), label='Up')
-            # plt.xlabel('Longitude (Degrees)')
-            # plt.ylabel('Log Change in Field-Aligned Vector')
-            # plt.title("Sensitivity of Field-Aligned Unit Vector")
-            # plt.legend()
-            # plt.tight_layout()
-            # plt.savefig('method_fa_diff_v_longitude.pdf' )
-            # plt.close()
-
-            # calculate mean and standard deviation and then plot those
-            plt.figure()
-            yerrx = np.nanstd(np.log10(mx[:,:-1]), axis=0)
-            yerry = np.nanstd(np.log10(my[:,:-1]), axis=0)
-            yerrz = np.nanstd(np.log10(mz[:,:-1]), axis=0)
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mx[:,:-1]), axis=0)),
-                            yerr=yerrx, label='East')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(my[:,:-1]), axis=0)),
-                            yerr=yerry, label='North')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mz[:,:-1]), axis=0)),
-                            yerr=yerrz, label='Up')
-            plt.xlabel('Longitude (Degrees)')
-            plt.ylabel('Log Change in Meridional Vector')
-            plt.title("Sensitivity of Meridional Unit Vector")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('method_mer_diff_v_longitude.pdf' )
-            plt.close()
-
-        except:
-            pass
-
-
-    def test_integrated_unit_vector_component_refheight_sensitivity_plots(self):
-        import matplotlib.pyplot as plt
-
-        p_lats, p_longs, p_alts = gen_plot_grid_fixed_alt(550.)
-        # data returned are the locations along each direction
-        # the full range of points obtained by iterating over all
-        # recasting alts into a more convenient form for later calculation
-        p_alts = [p_alts[0]]*len(p_longs)
-
-        # zonal vector components
-        # +1 on length of longitude array supports repeating first element
-        # shows nice periodicity on the plots
-        zvx = np.zeros((len(p_lats), len(p_longs)+1))
-        zvy = zvx.copy(); zvz = zvx.copy()
-        #  meridional vecrtor components
-        mx = zvx.copy(); my = zvx.copy(); mz = zvx.copy()
-        # field aligned, along B
-        bx = zvx.copy(); by = zvx.copy(); bz = zvx.copy()
-        date = datetime.datetime(2000,1,1)
-        # set up multi
-        if self.dc is not None:
-            import itertools
-            targets = itertools.cycle(dc.ids)
-            pending = []
-            for i,p_lat in enumerate(p_lats):
-                # iterate through target cyclicly and run commands
-                print (i, p_lat)
-                dview.targets = targets.next()
-                pending.append(dview.apply_async(pymv.calculate_integrated_mag_drift_unit_vectors_ecef,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        steps=None, max_steps=1000, step_size=10.,
-                                                                        ref_height=240.))
-                pending.append(dview.apply_async(pymv.calculate_integrated_mag_drift_unit_vectors_ecef,[p_lat]*len(p_longs), p_longs,
-                                                                        p_alts, [date]*len(p_longs),
-                                                                        steps=None, max_steps=1000, step_size=10.,
-                                                                        ref_height=0.))
-
-            for i,p_lat in enumerate(p_lats):
-                print ('collecting ', i, p_lat)
-                # collect output from first run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pending.pop(0).get()
-                zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # collect output from second run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pending.pop(0).get()
-                _a, _b, _c = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                zvx[i,:-1] = (zvx[i,:-1] - _a)/zvx[i,:-1]
-                zvy[i,:-1] = (zvy[i,:-1] - _b)/zvy[i,:-1]
-                zvz[i,:-1] = (zvz[i,:-1] - _c)/zvz[i,:-1]
-
-                _a, _b, _c  = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                bx[i,:-1] = (bx[i,:-1] - _a)/bx[i,:-1]
-                by[i,:-1] = (by[i,:-1] - _b)/by[i,:-1]
-                bz[i,:-1] = (bz[i,:-1] - _c)/bz[i,:-1]
-
-                _a, _b, _c = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                mx[i,:-1] = (mx[i,:-1] - _a)/mx[i,:-1]
-                my[i,:-1] = (my[i,:-1] - _b)/my[i,:-1]
-                mz[i,:-1] = (mz[i,:-1] - _c)/mz[i,:-1]
-
-        else:
-            for i,p_lat in enumerate(p_lats):
-                print (i, p_lat)
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pymv.calculate_integrated_mag_drift_unit_vectors_ecef([p_lat]*len(p_longs), p_longs,
-                                                                                        p_alts, [date]*len(p_longs),
-                                                                                        steps=None, max_steps=10000, step_size=10.,
-                                                                                        ref_height=240.)
-                zvx[i,:-1], zvy[i,:-1], zvz[i,:-1] = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                bx[i,:-1], by[i,:-1], bz[i,:-1] = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                mx[i,:-1], my[i,:-1], mz[i,:-1] = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-
-                # second run
-                tzx, tzy, tzz, tbx, tby, tbz, tmx, tmy, tmz = pymv.calculate_integrated_mag_drift_unit_vectors_ecef([p_lat]*len(p_longs), p_longs,
-                                                                                        p_alts, [date]*len(p_longs),
-                                                                                        steps=None, max_steps=10000, step_size=10.,
-                                                                                        ref_height=0.)
-                _a, _b, _c = pymv.ecef_to_enu_vector(tzx, tzy, tzz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                zvx[i,:-1] = (zvx[i,:-1] - _a)/zvx[i,:-1]
-                zvy[i,:-1] = (zvy[i,:-1] - _b)/zvy[i,:-1]
-                zvz[i,:-1] = (zvz[i,:-1] - _c)/zvz[i,:-1]
-
-                _a, _b, _c  = pymv.ecef_to_enu_vector(tbx, tby, tbz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                bx[i,:-1] = (bx[i,:-1] - _a)/bx[i,:-1]
-                by[i,:-1] = (by[i,:-1] - _b)/by[i,:-1]
-                bz[i,:-1] = (bz[i,:-1] - _c)/bz[i,:-1]
-
-                _a, _b, _c = pymv.ecef_to_enu_vector(tmx, tmy, tmz, [p_lat]*len(p_longs), p_longs)
-                # take difference with first run
-                mx[i,:-1] = (mx[i,:-1] - _a)/mx[i,:-1]
-                my[i,:-1] = (my[i,:-1] - _b)/my[i,:-1]
-                mz[i,:-1] = (mz[i,:-1] - _c)/mz[i,:-1]
-
-
-        # account for periodicity
-        zvx[:,-1] = zvx[:,0]
-        zvy[:,-1] = zvy[:,0]
-        zvz[:,-1] = zvz[:,0]
-        bx[:,-1] = bx[:,0]
-        by[:,-1] = by[:,0]
-        bz[:,-1] = bz[:,0]
-        mx[:,-1] = mx[:,0]
-        my[:,-1] = my[:,0]
-        mz[:,-1] = mz[:,0]
-
-        ytickarr = np.array([0, 0.25, 0.5, 0.75, 1])*(len(p_lats)-1)
-        xtickarr = np.array([0, 0.2, 0.4, 0.6, 0.8, 1])*len(p_longs)
-
-        try:
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Nornalized Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_east_diff_height.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvy)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Nornalized Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_north_diff_height.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(zvz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Zonal Unit Vector Nornalized Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_zonal_up_diff_height.pdf')
-            plt.close()
-
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(bx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Nornalized Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_east_diff_height.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(by)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Nornalized Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_north_diff_height.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(bz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Field Aligned Unit Vector Nornalized Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_fa_up_diff_height.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mx)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Nornalized Difference - Eastward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_east_diff_height.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(my)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Nornalized Difference - Northward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_north_diff_height.pdf')
-            plt.close()
-
-            fig = plt.figure()
-            plt.imshow(np.log10(np.abs(mz)), origin='lower')
-            plt.colorbar()
-            plt.yticks(ytickarr, ['-50', '-25', '0', '25', '50'])
-            plt.xticks(xtickarr, ['0', '72', '144', '216', '288', '360'])
-            plt.title('Log Meridional Unit Vector Nornalized Difference - Upward')
-            plt.xlabel('Geodetic Longitude (Degrees)')
-            plt.ylabel('Geodetic Latitude (Degrees)')
-            plt.savefig('int_mer_up_diff_height.pdf')
-            plt.close()
-
-            # calculate mean and standard deviation and then plot those
-            fig = plt.figure()
-            yerrx = np.nanstd(np.log10(zvx[:,:-1]), axis=0)
-            yerry = np.nanstd(np.log10(zvy[:,:-1]), axis=0)
-            yerrz = np.nanstd(np.log10(zvz[:,:-1]), axis=0)
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvx[:,:-1]), axis=0)),
-                            yerr=yerrx, label='East')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvy[:,:-1]), axis=0)),
-                            yerr=yerry, label='North')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(zvz[:,:-1]), axis=0)),
-                            yerr=yerrz, label='Up')
-            plt.xlabel('Longitude (Degrees)')
-            plt.ylabel('Log Normalized Change in Zonal Vector')
-            plt.title("Sensitivity of Zonal Unit Vector")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('int_zonal_diff_v_longitude_height.pdf' )
-            plt.close()
-
-            # calculate mean and standard deviation and then plot those
-            fig = plt.figure()
-            yerrx = np.nanstd(np.log10(mx[:,:-1]), axis=0)
-            yerry = np.nanstd(np.log10(my[:,:-1]), axis=0)
-            yerrz = np.nanstd(np.log10(mz[:,:-1]), axis=0)
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mx[:,:-1]), axis=0)),
-                            yerr=yerrx, label='East')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(my[:,:-1]), axis=0)),
-                            yerr=yerry, label='North')
-            plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(mz[:,:-1]), axis=0)),
-                            yerr=yerrz, label='Up')
-            plt.xlabel('Longitude (Degrees)')
-            plt.ylabel('Log Normalized Change in Meridional Vector')
-            plt.title("Sensitivity of Meridional Unit Vector")
-            plt.legend()
-            plt.tight_layout()
-            plt.savefig('int_mer_diff_v_longitude_height.pdf' )
-            plt.close()
-
-            # # calculate mean and standard deviation and then plot those
-            # fig = plt.figure()
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(bx[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(bx[:,:-1])), axis=0), label='East')
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(by[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(by[:,:-1])), axis=0), label='North')
-            # plt.errorbar(p_longs, np.log10(np.nanmedian(np.abs(bz[:,:-1]), axis=0)),
-            #                 yerr=np.nanstd(np.log10(np.abs(bz[:,:-1])), axis=0), label='Up')
-            # plt.xlabel('Longitude (Degrees)')
-            # plt.ylabel('Log Normalized Change in Field-Aligned Vector')
-            # plt.title("Sensitivity of Field-Aligned Unit Vector")
-            # plt.legend()
-            # plt.tight_layout()
-            # plt.savefig('int_fa_diff_v_longitude_height.pdf' )
-            # plt.close()
-
-        except:
-            pass
 
 
     def step_along_mag_unit_vector_sensitivity_plots(self, direction=None):
@@ -3990,6 +1477,8 @@ class TestCore():
         yield (f, )
         f = functools.partial(self.step_along_mag_unit_vector_sensitivity_plots, direction='meridional')
         yield (f, )
+
+
 
     def test_geomag_efield_scalars_plots(self):
         import matplotlib.pyplot as plt
@@ -4220,225 +1709,136 @@ class TestCore():
             pass
 
 
-    def test_basic_ecef_to_enu_rotations(self):
-        # test basic transformations first
-        # vector pointing along ecef y at 0, 0 is east
-        ve, vn, vu = pymv.ecef_to_enu_vector(0., 1., 0., 0., 0.)
-        # print ('{:9f}, {:9f}, {:9f}'.format(ve, vn, vu))
-        asseq(ve, 1.0, 9)
-        asseq(vn, 0, 9)
-        asseq(vu, 0, 9)
-        # vector pointing along ecef x at 0, 0 is up
-        ve, vn, vu = pymv.ecef_to_enu_vector(1., 0., 0., 0., 0.)
-        asseq(ve, 0.0, 9)
-        asseq(vn, 0.0, 9)
-        asseq(vu, 1.0, 9)
-        # vector pointing along ecef z at 0, 0 is north
-        ve, vn, vu = pymv.ecef_to_enu_vector(0., 0., 1., 0., 0.)
-        asseq(ve, 0.0, 9)
-        asseq(vn, 1.0, 9)
-        asseq(vu, 0.0, 9)
+    def test_unit_vector_and_field_line_plots(self):
+        import matplotlib.pyplot as plt
+        from mpl_toolkits.mplot3d import Axes3D
+        import os
+        on_travis = os.environ.get('ONTRAVIS') == 'True'
 
-        # vector pointing along ecef x at 0, 90 long is west
-        ve, vn, vu = pymv.ecef_to_enu_vector(1., 0., 0., 0., 90.)
-        # print ('{:9f}, {:9f}, {:9f}'.format(ve, vn, vu))
-        asseq(ve, -1.0, 9)
-        asseq(vn, 0.0, 9)
-        asseq(vu, 0.0, 9)
-        # vector pointing along ecef y at 0, 90 long is up
-        ve, vn, vu = pymv.ecef_to_enu_vector(0., 1., 0., 0., 90.)
-        asseq(ve, 0.0, 9)
-        asseq(vn, 0.0, 9)
-        asseq(vu, 1.0, 9)
-        # vector pointing along ecef z at 0, 90 long is north
-        ve, vn, vu = pymv.ecef_to_enu_vector(0., 0., 1., 0., 90.)
-        asseq(ve, 0.0, 9)
-        asseq(vn, 1.0, 9)
-        asseq(vu, 0.0, 9)
+        # convert OMNI position to ECEF
+        p_long = np.arange(0.,360.,12.)
+        p_alt = 0*p_long + 550.
+        p_lats = [ 5., 10., 15., 20., 25., 30.]
 
-        # vector pointing along ecef y at 0, 180 is west
-        ve, vn, vu = pymv.ecef_to_enu_vector(0., 1., 0., 0., 180.)
-        asseq(ve, -1.0, 9)
-        asseq(vn, 0.0, 9)
-        asseq(vu, 0.0, 9)
-        # vector pointing along ecef x at 0, 180 is down
-        ve, vn, vu = pymv.ecef_to_enu_vector(1., 0., 0., 0., 180.)
-        asseq(ve, 0.0, 9)
-        asseq(vn, 0.0, 9)
-        asseq(vu, -1.0, 9)
-        # vector pointing along ecef z at 0, 0 is north
-        ve, vn, vu = pymv.ecef_to_enu_vector(0., 0., 1., 0., 180.)
-        asseq(ve, 0.0, 9)
-        asseq(vn, 1.0, 9)
-        asseq(vu, 0.0, 9)
+        truthiness = []
+        for i,p_lat in enumerate(p_lats):
 
-        ve, vn, vu = pymv.ecef_to_enu_vector(0., 1., 0., 45., 0.)
-        # print ('{:9f}, {:9f}, {:9f}'.format(ve, vn, vu))
-        asseq(ve, 1.0, 9)
-        asseq(vn, 0, 9)
-        asseq(vu, 0, 9)
-        # vector pointing along ecef x at 0, 0 is south/up
-        ve, vn, vu = pymv.ecef_to_enu_vector(1., 0., 0., 45., 0.)
-        asseq(ve, 0.0, 9)
-        asseq(vn, -np.cos(np.pi/4), 9)
-        asseq(vu, np.cos(np.pi/4), 9)
-        # vector pointing along ecef z at 45, 0 is north/up
-        ve, vn, vu = pymv.ecef_to_enu_vector(0., 0., 1., 45., 0.)
-        asseq(ve, 0.0, 9)
-        asseq(vn, np.cos(np.pi/4), 9)
-        asseq(vu, np.cos(np.pi/4), 9)
+            trace_s = []
+            if not on_travis:
+                fig = plt.figure()
+                ax = fig.add_subplot(111, projection='3d')
 
+            #
+            date = datetime.datetime(2000,1,1)
+            ecef_x,ecef_y,ecef_z = pymv.geocentric_to_ecef(p_lat, p_long, p_alt)
 
-    def test_basic_enu_to_ecef_rotations(self):
-        # test basic transformations first
-        # vector pointing east at 0, 0 is along y
-        vx, vy, vz = pymv.enu_to_ecef_vector(1., 0., 0., 0., 0.)
-        # print ('{:9f}, {:9f}, {:9f}'.format(vx, vy, vz))
-        asseq(vx, 0.0, 9)
-        asseq(vy, 1.0, 9)
-        asseq(vz, 0.0, 9)
-        # vector pointing up at 0, 0 is along x
-        vx, vy, vz = pymv.enu_to_ecef_vector(0., 0., 1., 0., 0.)
-        asseq(vx, 1.0, 9)
-        asseq(vy, 0.0, 9)
-        asseq(vz, 0.0, 9)
-        # vector pointing north at 0, 0 is along z
-        vx, vy, vz = pymv.enu_to_ecef_vector(0., 1., 0., 0., 0.)
-        asseq(vx, 0.0, 9)
-        asseq(vy, 0.0, 9)
-        asseq(vz, 1.0, 9)
+            for j,(x,y,z) in enumerate(zip(ecef_x, ecef_y, ecef_z)):
+                # perform field line traces
+                trace_n = pymv.field_line_trace(np.array([x,y,z]), date, 1., 0., step_size=.5, max_steps=1.E6)
+                trace_s = pymv.field_line_trace(np.array([x,y,z]), date, -1., 0., step_size=.5, max_steps=1.E6)
+                # combine together, S/C position is first for both
+                # reverse first array and join so plotting makes sense
+                trace = np.vstack((trace_n[::-1], trace_s))
+                trace = pds.DataFrame(trace, columns=['x','y','z'])
+                # plot field-line
+                if not on_travis:
+                    ax.plot(trace['x'],trace['y'],trace['z'] , 'b')
+                    plt.xlabel('X')
+                    plt.ylabel('Y')
+                    ax.set_zlabel('Z')
+                # clear stored data
+                self.inst.data = pds.DataFrame()
+                # downselect, reduce number of points
+                trace = trace.ix[::1000,:]
 
-        # east vector at 0, 90 long points along -x
-        vx, vy, vz = pymv.enu_to_ecef_vector(1., 0., 0., 0., 90.)
-        # print ('{:9f}, {:9f}, {:9f}'.format(vx, vy, vz))
-        asseq(vx, -1.0, 9)
-        asseq(vy, 0.0, 9)
-        asseq(vz, 0.0, 9)
-        # vector pointing up at 0, 90 is along y
-        vx, vy, vz = pymv.enu_to_ecef_vector(0., 0., 1., 0., 90.)
-        asseq(vx, 0.0, 9)
-        asseq(vy, 1.0, 9)
-        asseq(vz, 0.0, 9)
-        # vector pointing north at 0, 90 is along z
-        vx, vy, vz = pymv.enu_to_ecef_vector(0., 1., 0., 0., 90.)
-        asseq(vx, 0.0, 9)
-        asseq(vy, 0.0, 9)
-        asseq(vz, 1.0, 9)
+                # compute magnetic field vectors
+                # need to provide alt, latitude, and longitude in geodetic coords
+                latitude, longitude, altitude = pymv.ecef_to_geodetic(trace['x'], trace['y'], trace['z'])
+                self.inst[:,'latitude'] = latitude
+                self.inst[:,'longitude'] = longitude
+                self.inst[:,'altitude'] = altitude
+                # store values for plotting locations for vectors
+                self.inst[:,'x'] = trace['x'].values
+                self.inst[:,'y'] = trace['y'].values
+                self.inst[:,'z'] = trace['z'].values
+                self.inst.data = self.inst[self.inst['altitude'] > 250.]
 
-        # vector pointing east at 0, 0 is along y
-        vx, vy, vz = pymv.enu_to_ecef_vector(1., 0., 0., 0., 180.)
-        # print ('{:9f}, {:9f}, {:9f}'.format(vx, vy, vz))
-        asseq(vx, 0.0, 9)
-        asseq(vy, -1.0, 9)
-        asseq(vz, 0.0, 9)
-        # vector pointing up at 0, 180 is along -x
-        vx, vy, vz = pymv.enu_to_ecef_vector(0., 0., 1., 0., 180.)
-        asseq(vx, -1.0, 9)
-        asseq(vy, 0.0, 9)
-        asseq(vz, 0.0, 9)
-        # vector pointing north at 0, 180 is along z
-        vx, vy, vz = pymv.enu_to_ecef_vector(0., 1., 0., 0., 180.)
-        asseq(vx, 0.0, 9)
-        asseq(vy, 0.0, 9)
-        asseq(vz, 1.0, 9)
+                # also need to provide transformation from ECEF to S/C
+                # going to leave that a null transformation so we can plot in ECF
+                self.inst[:,'sc_xhat_x'], self.inst[:,'sc_xhat_y'], self.inst[:,'sc_xhat_z'] = 1., 0., 0.
+                self.inst[:,'sc_yhat_x'], self.inst[:,'sc_yhat_y'], self.inst[:,'sc_yhat_z'] = 0., 1., 0.
+                self.inst[:,'sc_zhat_x'], self.inst[:,'sc_zhat_y'], self.inst[:,'sc_zhat_z'] = 0., 0., 1.
+                self.inst.data.index = pysat.utils.time.create_date_range(pysat.datetime(2000,1,1),
+                                                                    pysat.datetime(2000,1,1)+pds.DateOffset(seconds=len(self.inst.data)-1),
+                                                                    freq='S')
+                pymv.satellite.add_mag_drift_unit_vectors(self.inst)
 
-    def test_ecef_to_enu_back_to_ecef(self):
+                #if i % 2 == 0:
+                length = 500
+                vx = self.inst['unit_zon_x']
+                vy = self.inst['unit_zon_y']
+                vz = self.inst['unit_zon_z']
+                if not on_travis:
+                    ax.quiver3D(self.inst['x'] + length*vx, self.inst['y'] + length*vy,
+                                self.inst['z'] + length*vz, vx, vy,vz, length=500.,
+                                color='green') #, pivot='tail')
+                length = 500
+                vx = self.inst['unit_fa_x']
+                vy = self.inst['unit_fa_y']
+                vz = self.inst['unit_fa_z']
+                if not on_travis:
+                    ax.quiver3D(self.inst['x'] + length*vx, self.inst['y'] + length*vy,
+                                self.inst['z'] + length*vz, vx, vy,vz, length=500.,
+                                color='purple') #, pivot='tail')
+                length = 500
+                vx = self.inst['unit_mer_x']
+                vy = self.inst['unit_mer_y']
+                vz = self.inst['unit_mer_z']
+                if not on_travis:
+                    ax.quiver3D(self.inst['x'] + length*vx, self.inst['y'] + length*vy,
+                                self.inst['z'] + length*vz, vx, vy,vz, length=500.,
+                                color='red') #, pivot='tail')
 
-        vx = 0.9
-        vy = 0.1
-        vz = np.sqrt(1. - vx**2+vy**2)
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        for lat, lon, alt in zip(lats, longs, alts):
-            vxx, vyy, vzz = pymv.ecef_to_enu_vector(vx, vy, vz, lat, lon)
-            vxx, vyy, vzz = pymv.enu_to_ecef_vector(vxx, vyy, vzz, lat, lon)
-            asseq(vx, vxx, 9)
-            asseq(vy, vyy, 9)
-            asseq(vz, vzz, 9)
+                # check that vectors norm to 1
+                assert np.all(np.sqrt(self.inst['unit_zon_x']**2 + self.inst['unit_zon_y']**2 + self.inst['unit_zon_z']**2) > 0.999999)
+                assert np.all(np.sqrt(self.inst['unit_fa_x']**2 + self.inst['unit_fa_y']**2 + self.inst['unit_fa_z']**2) > 0.999999)
+                assert np.all(np.sqrt(self.inst['unit_mer_x']**2 + self.inst['unit_mer_y']**2 + self.inst['unit_mer_z']**2) > 0.999999)
+                # confirm vectors are mutually orthogonal
+                dot1 =  self.inst['unit_zon_x']*self.inst['unit_fa_x'] + self.inst['unit_zon_y']*self.inst['unit_fa_y']  + self.inst['unit_zon_z']*self.inst['unit_fa_z']
+                dot2 =  self.inst['unit_zon_x']*self.inst['unit_mer_x'] + self.inst['unit_zon_y']*self.inst['unit_mer_y']  + self.inst['unit_zon_z']*self.inst['unit_mer_z']
+                dot3 =  self.inst['unit_fa_x']*self.inst['unit_mer_x'] + self.inst['unit_fa_y']*self.inst['unit_mer_y']  + self.inst['unit_fa_z']*self.inst['unit_mer_z']
+                assert np.all(np.abs(dot1) < 1.E-6)
+                assert np.all(np.abs(dot2) < 1.E-6)
+                assert np.all(np.abs(dot3) < 1.E-6)
 
-    def test_enu_to_ecef_back_to_enu(self):
+                # ensure that zonal vector is generally eastward
+                ones = np.ones(len(self.inst.data.index))
+                zeros = np.zeros(len(self.inst.data.index))
+                ex, ey, ez = pymv.enu_to_ecef_vector(ones, zeros, zeros, self.inst['latitude'], self.inst['longitude'])
+                nx, ny, nz = pymv.enu_to_ecef_vector(zeros, ones, zeros, self.inst['latitude'], self.inst['longitude'])
+                ux, uy, uz = pymv.enu_to_ecef_vector(zeros, zeros, ones, self.inst['latitude'], self.inst['longitude'])
 
-        vx = 0.9
-        vy = 0.1
-        vz = np.sqrt(1. - vx**2+vy**2)
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        for lat, lon, alt in zip(lats, longs, alts):
-            vxx, vyy, vzz = pymv.enu_to_ecef_vector(vx, vy, vz, lat, lon)
-            vxx, vyy, vzz = pymv.ecef_to_enu_vector(vxx, vyy, vzz, lat, lon)
-            asseq(vx, vxx, 9)
-            asseq(vy, vyy, 9)
-            asseq(vz, vzz, 9)
+                dot1 =  self.inst['unit_zon_x']*ex + self.inst['unit_zon_y']*ey  + self.inst['unit_zon_z']*ez
+                assert np.all(dot1 > 0.)
 
-    def test_igrf_end_to_ecef_back_to_end(self):
-        #import pdb
-        vx = 0.9
-        vy = 0.1
-        vz = np.sqrt(1. - vx**2+vy**2)
-        vz = -vz
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        for lat, lon, alt in zip(lats, longs, alts):
-            #print(vx, vy, vz, lat, lon)
-            #pdb.set_trace()
-            # input here is co-latitude, not latitude
-            # inputs to fortran are in radians
-            vxx, vyy, vzz = igrf.end_vector_to_ecef(vx, vy, vz, np.deg2rad(90.-lat), np.deg2rad(lon))
-            vx2, vy2, vz2 = pymv.enu_to_ecef_vector(vx, vy, -vz, lat, lon)
-            # print ('end check ', vxx, vyy, vzz, vx2, vy2, vz2)
-            asseq(vxx, vx2, 9)
-            asseq(vyy, vy2, 9)
-            asseq(vzz, vz2, 9)
+                dot1 =  self.inst['unit_fa_x']*nx + self.inst['unit_fa_y']*ny  + self.inst['unit_fa_z']*nz
+                assert np.all(dot1 > 0.)
 
-            vxx, vyy, vzz = pymv.ecef_to_enu_vector(vxx, vyy, vzz, lat, lon)
-            # convert upward component back to down
-            vzz = -vzz
-            # compare original inputs to outputs
-            asseq(vx, vxx, 9)
-            asseq(vy, vyy, 9)
-            asseq(vz, vzz, 9)
+                dot1 =  self.inst['unit_mer_x']*ux + self.inst['unit_mer_y']*uy  + self.inst['unit_mer_z']*uz
+                assert np.all(dot1 > 0.)
 
 
-    def test_igrf_ecef_to_geodetic_back_to_ecef(self):
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        for ecef_x, ecef_y, ecef_z, geo_lat, geo_lon, geo_alt in zip(ecf_x,ecf_y,
-                           ecf_z, lats, longs, alts):
-            pos = np.array([ecef_x, ecef_y, ecef_z])
-            lat, elong, alt = igrf.ecef_to_geodetic(pos)
-            lat = np.rad2deg(lat)
-            elong = np.rad2deg(elong)
-            if (elong < 0):
-                elong = elong + 360.
+            if not on_travis:
+                plt.savefig(''.join(('magnetic_unit_vectors_',str(int(p_lat)),'.pdf')))
+                plt.close()
 
-            d_lat = lat - geo_lat
-            d_long = elong - geo_lon
-            d_alt = alt - geo_alt
-
-            # print ('Word', ecef_x, ecef_y, ecef_z)
-            # print (geo_lat, geo_lon, geo_alt)
-            # print (lat, elong, alt)
-            assert np.all(np.abs(d_lat) < 1.E-5)
-            assert np.all(np.abs(d_long) < 1.E-5)
-            assert np.all(np.abs(d_alt) < 1.E-5)
-
-
-    def test_igrf_ecef_to_geographic_with_colatitude(self):
-        lats, longs, alts = gen_data_fixed_alt(550.)
-        ecf_x,ecf_y,ecf_z = pymv.geodetic_to_ecef(lats,
-                                                  longs,
-                                                  alts)
-        for ecef_x, ecef_y, ecef_z, geo_lat, geo_lon, geo_alt in zip(ecf_x,ecf_y,
-                           ecf_z, lats, longs, alts):
-            pos = np.array([ecef_x, ecef_y, ecef_z])
-
-            colat, lon, r = igrf.ecef_to_colat_long_r(pos)
-            # results are returned in radians
-            lat = 90. - np.rad2deg(colat)
-            lon = np.rad2deg(lon)
-
-            lat2, lon2, h2 = pymv.ecef_to_geocentric(*pos, ref_height=0)
-
-            # print(lat, lon, r, lat2, lon2, h2)
-            asseq(r, h2, 9)
-            asseq(lat, lat2, 9)
-            asseq(lon, lon2, 9)
+        ## plot Earth
+        #u = np.linspace(0, 2 * np.pi, 100)
+        #v = np.linspace(60.*np.pi/180., 120.*np.pi/180., 100)
+        #xx = 6371. * np.outer(np.cos(u), np.sin(v))
+        #yy = 6371. * np.outer(np.sin(u), np.sin(v))
+        #zz = 6371. * np.outer(np.ones(np.size(u)), np.cos(v))
+        #ax.plot_surface(xx, yy, zz, rstride=4, cstride=4, color='darkgray')
+        #plt.savefig('magnetic_unit_vectors_w_globe.pdf')
+        #print truthiness
+        assert np.all(truthiness)
